Split the two files that were holding a third of the package
store.go was 2,555 lines and service.go 1,653 — between them a fifth of
all non-test Go, and the two files anybody changing behavior has to open.
Neither had drifted for a reason: docs/surface.md deliberately does not
count implementation lines, so nothing was ever going to say they had
grown, and every addition landed at the bottom of whichever file already
had a function nearby.

The seams were already there in the prose. store.go now keeps the
connection, the columns, the scan helpers and a concept's lifecycle;
what came out is what a reader was scrolling past to reach it — filter
(the WHERE clause every read shares), search (fragmenting for pg_trgm and
the two vector paths), list (the keyset feeds), move (the one operation
that writes rows it was not asked about), revision, embedding. service.go
keeps the write path and its policy, with search, get_context, listings,
the loop's measurements and the embedding backfill beside it.

Nothing else changed. Every line is where it was, in a different file:
the decls were cut with their doc comments and reassembled, and the
multiset of non-import lines is identical to what it was before. The new
files carry a sentence each about what they hold, since the package
comment now only belongs on one of them.

No counter is added for this, deliberately. docs/surface.md says what it
leaves out and why — implementation lines are not something a user pays
for, and the trouble they cause is the maintainer's. That is the honest
position, and it is also why this file will grow again unless somebody
opens it.
