+++ internal/service/service.go
@@ -5,21 +5,14 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log/slog"
-	"maps"
-	"slices"
-	"sort"
 	"strings"
-	"time"
-	"unicode/utf8"
 
 	"github.com/na0fu3y/ochakai/internal/config"
 	"github.com/na0fu3y/ochakai/internal/domain"
 	"github.com/na0fu3y/ochakai/internal/embed"
-	"github.com/na0fu3y/ochakai/internal/httpauth"
 	"github.com/na0fu3y/ochakai/internal/okf"
 	"github.com/na0fu3y/ochakai/internal/store"
 )
@@ -684,970 +677,3 @@ func validateOKF(k *domain.Knowledge) error {
 }
 
 // --- search ---
-
-// Search runs lexical search, and when an embedder is configured, fuses
-// it with vector search via reciprocal rank fusion.
-func (s *Service) Search(ctx context.Context, query string, f store.Filter, limit int) ([]domain.SearchHit, error) {
-	// Stored text is NFC (design doc 0022); an NFD query (pasted from a
-	// macOS path) must still match it byte-wise.
-	query = domain.Normalize(query)
-	// An empty query has nothing to rank by: SearchLexical splits the
-	// query into fragments and gets none, so guard here — once, for
-	// every surface — and point the caller at the listing modes.
-	if strings.TrimSpace(query) == "" {
-		// The modes come from domain.ListSorts rather than a sentence:
-		// this message named three of them for as long as there were
-		// three, and stale_after (design doc 0037) never reached it.
-		return nil, Invalidf("search needs a query; use sort=%s to list concepts without one, "+
-			"or source=URI to list what cites a resource", strings.Join(domain.ListSorts, "|"))
-	}
-	f, err := checkedFilter(f)
-	if err != nil {
-		return nil, err
-	}
-	hits, err := s.search(ctx, query, f, limit)
-	if err != nil {
-		return nil, err
-	}
-	ids := make([]string, len(hits))
-	for i, h := range hits {
-		ids[i] = h.ID
-	}
-	s.recordUsage(ctx, domain.EventSearchHit, ids)
-	// A search that found nothing is recorded as itself: it is the one
-	// observation that says what this knowledge base is missing, and
-	// until design doc 0051 it was the only one being discarded. Here
-	// rather than in each surface, so every way of asking is counted —
-	// get_context searches through this function too.
-	if len(hits) == 0 {
-		s.recordMiss(ctx, query)
-	}
-	return hits, nil
-}
-
-func (s *Service) search(ctx context.Context, query string, f store.Filter, limit int) ([]domain.SearchHit, error) {
-	if limit <= 0 || limit > 50 {
-		limit = 10
-	}
-	lexical, err := s.Store.SearchLexical(ctx, query, f, limit*2)
-	if err != nil {
-		return nil, err
-	}
-	if s.Embedder == nil {
-		if len(lexical) > limit {
-			lexical = lexical[:limit]
-		}
-		return lexical, nil
-	}
-
-	vecs, err := s.Embedder.Embed(ctx, embed.TaskQuery, []string{query})
-	if err != nil {
-		// Degrade to lexical-only rather than failing the search.
-		s.Log.Warn("query embedding failed; falling back to lexical-only", "error", err)
-		if len(lexical) > limit {
-			lexical = lexical[:limit]
-		}
-		return lexical, nil
-	}
-	vector, err := s.Store.SearchVector(ctx, vecs[0], s.Embedder.Model(), f, limit*2)
-	if err != nil {
-		return nil, err
-	}
-	// Concepts whose attachments match are the third list (design doc
-	// 0020): a concept matching in both body and attachment gains rank
-	// from both, so evidence-backed concepts surface first.
-	attachments, err := s.Store.SearchVectorAttachments(ctx, vecs[0], s.Embedder.Model(), f, limit*2)
-	if err != nil {
-		return nil, err
-	}
-	fused := rrfFuse(limit, lexical, vector, attachments)
-	return fused, nil
-}
-
-// ContextResult is the one-call context pack behind get_context and
-// `ochakai context`: the ranking, plus the full concepts behind the top
-// hits expanded one hop through links. Concepts that did not fit the
-// caller's budget are listed in Outline instead — named, not delivered.
-//
-// Hits is the ranking only (design doc 0033). The knowledge travels once,
-// in Concepts; a hit that was not expanded into a concept is a pointer the
-// caller can spend a round trip on.
-type ContextResult struct {
-	Hits      []domain.ContextRank    `json:"hits"`
-	Entries   []domain.View           `json:"entries"`
-	Outline   []domain.ContextOutline `json:"outline,omitempty"`
-	Truncated int                     `json:"truncated,omitempty"`
-}
-
-// ContextRequest is the input to Context. Budget caps the serialized size
-// of the response — the concepts returned in full and the outline rows
-// naming the rest, which carry a description and are not free; 0 means no
-// cap.
-type ContextRequest struct {
-	Query  string
-	Filter store.Filter
-	Limit  int
-	Budget int
-}
-
-// Context gathers what an agent should read before answering a data
-// question, in one call: search (verified concepts rank higher), fetch
-// the concepts behind the top hits in full, and follow their links one
-// hop so companion knowledge — the insight that says how to read a
-// metric, the golden query that answers the question — arrives without
-// further round trips.
-//
-// req.Budget caps how many bytes of full concepts come back; the rest
-// become outline rows (packWithinBudget). Callers whose context window is
-// the scarce resource — an agent, not a UI — should always set it.
-func (s *Service) Context(ctx context.Context, req ContextRequest) (*ContextResult, error) {
-	if strings.TrimSpace(req.Query) == "" {
-		return nil, Invalidf("invalid context request: query is required")
-	}
-	limit := req.Limit
-	if limit <= 0 || limit > 20 {
-		limit = 5
-	}
-	hits, err := s.Search(ctx, req.Query, req.Filter, 2*limit)
-	if err != nil {
-		return nil, err
-	}
-	seen := map[string]bool{}
-	var entries []domain.Knowledge
-	addFetched := func(k *domain.Knowledge) {
-		if len(entries) >= 2*limit || seen[k.ID] || k.Rejection != nil {
-			return // rejected companions stay out of the pack
-		}
-		seen[k.ID] = true
-		entries = append(entries, *k)
-	}
-	add := func(id string) {
-		if len(entries) >= 2*limit || seen[id] {
-			return
-		}
-		k, err := s.Store.Get(ctx, id)
-		if err != nil {
-			return // deleted targets stay out of the pack
-		}
-		addFetched(k)
-	}
-	for _, h := range hits {
-		if len(entries) >= limit {
-			break
-		}
-		add(h.ID)
-	}
-	// One hop through the primary concepts' links, both directions: the
-	// query a metric links to, and the insight that links to the metric
-	// (rel: explains points at the metric, not the other way round).
-	// Companions share the 2*limit cap and are never expanded themselves.
-	primaries := len(entries)
-	for i := range primaries {
-		for _, l := range entries[i].Links {
-			add(l.Target)
-		}
-		linking, err := s.Store.ListLinkingToDocs(ctx, entries[i].ID, 2*limit)
-		if err != nil {
-			s.Log.Warn("backlink lookup failed", "id", entries[i].ID, "error", err)
-			continue
-		}
-		for j := range linking {
-			addFetched(&linking[j])
-		}
-	}
-	// Concepts become views here: a context pack hands over the document
-	// (design doc 0043 §3.5), and the budget below has to measure what is
-	// actually sent.
-	views := make([]domain.View, 0, len(entries))
-	for i := range entries {
-		v, err := okf.ViewOf(&entries[i])
-		if err != nil {
-			// The concept came from the store, so this means the renderer
-			// cannot express something it holds. Naming it and leaving it
-			// out beats failing the whole pack.
-			s.Log.Warn("cannot render a concept for a context pack", "id", entries[i].ID, "error", err)
-			continue
-		}
-		views = append(views, v)
-	}
-	kept, outline := packWithinBudget(views, req.Budget)
-	// Only delivered concepts count as fetched. An outline row names an
-	// concept; it does not hand over the knowledge, so counting it as a use
-	// would inflate the demand signal that drives the review feeds.
-	ids := make([]string, len(kept))
-	for i := range kept {
-		ids[i] = kept[i].ID
-	}
-	s.recordUsage(ctx, domain.EventFetched, ids)
-	return &ContextResult{
-		Hits: domain.ContextRanks(hits), Entries: kept,
-		Outline: outline, Truncated: len(outline),
-	}, nil
-}
-
-// packWithinBudget splits concepts into the ones that fit within budget
-// serialized bytes and outline rows for the rest, preserving rank order in
-// both. budget <= 0 means no cap.
-//
-// Concepts are atomic: a concept is delivered whole or not at all. Cutting a
-// body mid-way would be worse than dropping it — half of an attested
-// computation's SQL still looks executable, and half a model spec is not a
-// spec. Nothing downstream can tell a truncated field from a short one.
-//
-// Packing is greedy rather than a prefix cut: one oversized concept high in
-// the ranking (a model concept carries its entire spec in attrs) would
-// otherwise starve everything below it. A concept larger than the whole
-// budget always becomes an outline row, even at rank 1 — the caller sees
-// its size and can fetch it deliberately.
-//
-// The outline is inside the budget, not on top of it. Naming what was left
-// out costs bytes too, and an outline row carries a description, which is
-// unbounded on a concept: a budget that governed only the delivered concepts
-// left the response as a whole ungoverned, which is the thing callers
-// actually have to fit in a context window. So rows are counted, their
-// descriptions are capped, and concepts are demoted from the bottom of the
-// ranking until the two halves fit together.
-//
-// The floor is the outline of everything: a request whose budget cannot
-// even hold the names of what matched gets those names anyway. Silently
-// dropping concepts the caller never hears about is worse than a response
-// slightly over budget, and the caller can raise the budget only if it
-// knows there was something there.
-func packWithinBudget(entries []domain.View, budget int) ([]domain.View, []domain.ContextOutline) {
-	if budget <= 0 {
-		return entries, nil
-	}
-	sizes := make([]int, len(entries))
-	rows := make([]domain.ContextOutline, len(entries))
-	rowSizes := make([]int, len(entries))
-	deliver := make([]bool, len(entries))
-	for i := range entries {
-		sizes[i] = serializedSize(&entries[i])
-		rows[i] = outlineRow(&entries[i], sizes[i])
-		rowSizes[i] = jsonSize(rows[i])
-	}
-	used, outlineBytes := 0, 0
-	for i := range entries {
-		if used+sizes[i] <= budget {
-			deliver[i] = true
-			used += sizes[i]
-			continue
-		}
-		outlineBytes += rowSizes[i]
-	}
-	// Demote from the bottom of the ranking until the outline the skipped
-	// concepts cost fits alongside what was kept. Each demotion frees the
-	// concept's bytes and pays for its row, which is strictly smaller — the
-	// guard is there so a pathological concept cannot spin this loop.
-	for used+outlineBytes > budget {
-		last := -1
-		for i := range deliver {
-			if deliver[i] {
-				last = i
-			}
-		}
-		if last < 0 || sizes[last] <= rowSizes[last] {
-			break
-		}
-		deliver[last] = false
-		used -= sizes[last]
-		outlineBytes += rowSizes[last]
-	}
-	kept := make([]domain.View, 0, len(entries))
-	var outline []domain.ContextOutline
-	for i := range entries {
-		if deliver[i] {
-			kept = append(kept, entries[i])
-			continue
-		}
-		outline = append(outline, rows[i])
-	}
-	return kept, outline
-}
-
-// outlineDescriptionBytes caps the description an outline row carries.
-// The description is what makes a row worth reading — it is the caller's
-// only basis for spending a round trip — but nothing bounds it on the
-// concept, so one concept with a page-long description could outweigh the
-// budget on its own. Enough for a sentence or two; the rest is what the
-// fetch is for.
-const outlineDescriptionBytes = 200
-
-func outlineRow(v *domain.View, size int) domain.ContextOutline {
-	return domain.ContextOutline{
-		ID: v.Summary.ID, Type: v.Summary.Type, Title: v.Summary.Title,
-		Description: truncateUTF8(v.Summary.Description, outlineDescriptionBytes),
-		Status:      v.Summary.Status, Bytes: size,
-	}
-}
-
-// serializedSize is what the concept costs on the wire. It measures the
-// whole view, document included: the document is the concept, so a measure
-// that skipped it would be budgeting for a fraction of what is sent.
-func serializedSize(v *domain.View) int {
-	b, err := json.Marshal(v)
-	if err != nil {
-		return len(v.Document) // unreachable in practice; never drop on a marshal quirk
-	}
-	return len(b)
-}
-
-// jsonSize is what any part of the response costs on the wire.
-func jsonSize(v any) int {
-	b, err := json.Marshal(v)
-	if err != nil {
-		return 0
-	}
-	return len(b)
-}
-
-// SearchOrList answers the query surface REST (GET /api/v1/search) and
-// MCP (search_concepts) both expose, so the two cannot drift: one place
-// decides whether a request searches or lists, validates the mode, and
-// runs it. A sort mode names one of the listing feeds (see list); with no
-// sort, a source filter and no query is the reverse lookup; otherwise it
-// is a search.
-//
-// Sort and query are mutually exclusive rather than combined: a feed is a
-// queue a reviewer works through, and ranking it by relevance to a query
-// would leave the reviewer with neither the whole queue nor the best
-// matches. The CLI refuses the same pair before it ever gets here.
-//
-// cursor resumes a listing where the previous page ended, and is refused
-// on a search: a listing has a total order to resume from, a ranking has
-// a window (design doc 0050 §2.2).
-func (s *Service) SearchOrList(ctx context.Context, query, sort, cursor string, f store.Filter, limit int) (*Listing, error) {
-	if sort != "" {
-		if !domain.ValidListSort(sort) {
-			return nil, Invalidf("invalid sort %q (valid: %s)", sort, strings.Join(domain.ListSorts, ", "))
-		}
-		if strings.TrimSpace(query) != "" {
-			return nil, Invalidf("sort=%s lists concepts; it cannot be combined with a search query", sort)
-		}
-		return s.list(ctx, sort, cursor, f, limit)
-	}
-	// A reverse lookup with nothing to search by lists rather than
-	// searches: "what derives from this material" (design doc 0037 §2.3)
-	// and "what links at this concept" both have a set for an answer and no
-	// text to rank it by. Both list in address order, and the mode is
-	// named so a cursor from one cannot resume the other.
-	if strings.TrimSpace(query) == "" {
-		switch {
-		case f.Source != "":
-			return s.list(ctx, sortBySource, cursor, f, limit)
-		case f.LinksTo != "":
-			return s.list(ctx, sortByLinksTo, cursor, f, limit)
-		}
-	}
-	if cursor != "" {
-		return nil, searchBoundError()
-	}
-	hits, err := s.Search(ctx, query, f, limit)
-	if err != nil {
-		return nil, err
-	}
-	return &Listing{Hits: hits}, nil
-}
-
-// The listing modes a reverse-lookup filter with no query selects.
-// Neither is one of domain.ListSorts: no caller can ask for them by name,
-// because the filter already says what is wanted. They are distinct so
-// that a cursor names the listing it came from.
-const (
-	sortBySource  = "source"
-	sortByLinksTo = "links_to"
-)
-
-// askableKeys are the frontmatter keys "fm." answers, and askableHint is
-// the list a refusal shows. Both come from the domain vocabulary, so the
-// server, the OpenAPI text, the MCP schemas and the CLI help cannot
-// disagree about what is askable.
-var (
-	askableKeys = domain.AskableFrontmatterKeys()
-	askableHint = strings.Join(askableKeys, ", ")
-)
-
-// checkedFilter returns f with its path scopes normalized, rejecting one
-// that could not lead an id, and refuses a frontmatter filter naming a key
-// "fm." does not answer. Every surface hands its filter through here — the
-// contract lives on the server (design doc 0015 §2), so REST, MCP and the
-// CLI cannot disagree about whether "teams/growth/" is the same scope as
-// "teams/growth", nor about which spellings of a question exist.
-//
-// An empty prefix is the root, which narrows nothing: it drops out rather
-// than erroring, because asking for everything is a request the server can
-// answer, not a mistake to report.
-//
-// This is not an access check. A caller may pass any prefix or none, and
-// none returns the whole base — reading is bounded by who can reach the
-// service, not by what they ask for (design doc 0002, and 0041 §4).
-func checkedFilter(f store.Filter) (store.Filter, error) {
-	// Sorted, so a request naming two bad keys reports the same one every
-	// time rather than whichever the map handed over first.
-	for _, key := range slices.Sorted(maps.Keys(f.Frontmatter)) {
-		if use, ok := domain.FilterOwnedKeys[key]; ok {
-			return f, Invalidf("fm.%s is not a filter: ochakai reads %s through a column of its own, "+
-				"which does not answer what the frontmatter says — use %s", key, key, use)
-		}
-		if !slices.Contains(askableKeys, key) {
-			return f, Invalidf("fm.%s is not a filter: OKF does not define %q, and the filter "+
-				"vocabulary is OKF's — a producer's own key is stored and handed back exactly as "+
-				"written, not asked for. Askable: %s", key, key, askableHint)
-		}
-	}
-	if len(f.Prefixes) == 0 {
-		return f, nil
-	}
-	scoped := make([]string, 0, len(f.Prefixes))
-	for _, p := range f.Prefixes {
-		clean, err := normalizePrefix(p)
-		if err != nil {
-			return f, err
-		}
-		if clean != "" {
-			scoped = append(scoped, clean)
-		}
-	}
-	f.Prefixes = scoped
-	return f, nil
-}
-
-// list runs one of the listing feeds. None of them is a search: no usage
-// is recorded (reading a review queue must not inflate the very signal it
-// ranks by), and every hit carries score 0, so the REST and MCP responses
-// keep the exact wire shape of a search across all modes.
-//
-//   - verified_at — by verification age, oldest first (never-verified
-//     last): the feed for canary runs over verified golden queries (see
-//     docs/guides/golden-query-canary.md).
-//   - usage — by demand, most-searched first, never-used drafts
-//     oldest-first at the bottom: the web UI's draft review queue. Hits
-//     carry their usage totals so the reviewer sees the evidence inline.
-//   - failed — concepts whose failure reports are still unanswered, worst
-//     first: the re-verification feed (design doc 0025), the
-//     evidence-based counterpart to verified_at. Verifying (or rejecting)
-//     takes a concept out of it, so a base that is kept up yields an empty
-//     one. Hits carry their usage totals.
-//   - stale_after — concepts whose declared expiry has passed, most
-//     overdue first (design doc 0037 §2.1); only concepts stale today, a
-//     date that has not come due is not work yet. The one feed verifying
-//     does not empty: verified_at and failed are server observations,
-//     while stale_after is what the writer declared, so clearing it means
-//     editing the concept to declare a new expiry (design doc 0037 §2.2).
-//   - source — the concepts citing one resource, the reverse of
-//     sources[].resource (design doc 0037 §2.3).
-//
-// Every mode pages: each has a total order ending in the id, so a cursor
-// carries a position in it and the caller walks past the cap instead of
-// stopping at it (design doc 0050 §2.1). One row beyond the page is read
-// to tell a full page from the last one — without it, a listing whose
-// length happens to equal the limit would hand out a cursor onto nothing.
-func (s *Service) list(ctx context.Context, sort, cursor string, f store.Filter, limit int) (*Listing, error) {
-	if limit <= 0 || limit > 1000 {
-		limit = 100
-	}
-	f, err := checkedFilter(f)
-	if err != nil {
-		return nil, err
-	}
-	after, err := decodeCursor(cursor, sort, cursorSortKeys(sort))
-	if err != nil {
-		return nil, err
-	}
-	hits, err := s.listPage(ctx, sort, f, after, limit+1)
-	if err != nil {
-		return nil, err
-	}
-	more := len(hits) > limit
-	if more {
-		hits = hits[:limit]
-	}
-	return &Listing{Hits: hits, Cursor: nextCursor(sort, hits, more)}, nil
-}
-
-// listPage runs one feed's query. The two usage feeds rank by the usage
-// totals and hand them back with each hit, so they are hits already; the
-// rest return concepts to project.
-func (s *Service) listPage(ctx context.Context, sort string, f store.Filter, after *store.After, limit int) ([]domain.SearchHit, error) {
-	switch sort {
-	case "usage":
-		return s.Store.ListByUsage(ctx, f, after, limit)
-	case "failed":
-		return s.Store.ListByFailed(ctx, f, after, limit)
-	}
-	list := s.Store.ListByVerifiedAt
-	switch sort {
-	case "stale_after":
-		list = s.Store.ListByStaleAfter
-	case sortBySource, sortByLinksTo:
-		// Both reverse lookups are the same query with a different
-		// filter, in address order (design doc 0050: a listing pages).
-		list = s.Store.ListByAddress
-	}
-	entries, err := list(ctx, f, after, limit)
-	if err != nil {
-		return nil, err
-	}
-	hits := make([]domain.SearchHit, len(entries))
-	for i := range entries {
-		hits[i] = domain.SearchHit{Summary: domain.SummaryOf(&entries[i])}
-	}
-	return hits, nil
-}
-
-// Revisions returns a concept's change history, newest first — the
-// audit surface behind "every change kept as a revision". Not a search:
-// no usage is recorded (auditing a concept is not using it).
-func (s *Service) Revisions(ctx context.Context, id string, limit int) ([]domain.Revision, error) {
-	if limit <= 0 || limit > 200 {
-		limit = 50
-	}
-	return s.Store.ListRevisions(ctx, domain.Normalize(id), limit)
-}
-
-// Usage returns usage totals for one concept (404 when the concept is gone).
-func (s *Service) Usage(ctx context.Context, id string) (*domain.Usage, error) {
-	id = domain.Normalize(id)
-	if _, err := s.Store.Get(ctx, id); err != nil {
-		return nil, err
-	}
-	return s.Store.Usage(ctx, id)
-}
-
-// maxOutcomeNote bounds the free-form note recorded with an outcome
-// report; notes live in raw knowledge_event rows (pruned after 180 days).
-const maxOutcomeNote = 2000
-
-// maxRejectionNote bounds a rejection's reason, matching the outcome
-// note: both are a human sentence or two explaining a judgment, and
-// neither is a place to paste a transcript.
-const maxRejectionNote = 2000
-
-// ReportOutcome records a worked/failed report against one concept and
-// returns the updated usage totals. The last edge of the write-back
-// loop: an agent that ran a golden query and got a wrong number can say
-// so, instead of the next agent trusting the same concept blind. Unlike
-// passive usage recording, a failed write is returned — the reporter
-// should know the report was lost.
-func (s *Service) ReportOutcome(ctx context.Context, id, outcome, note string) (*domain.Usage, error) {
-	if err := s.readOnly(); err != nil {
-		return nil, err
-	}
-	id = domain.Normalize(id)
-	if !domain.ValidOutcome(outcome) {
-		return nil, Invalidf("invalid outcome %q (valid: %s)", outcome, strings.Join(domain.Outcomes, ", "))
-	}
-	if len(note) > maxOutcomeNote {
-		return nil, Invalidf("note exceeds %d bytes", maxOutcomeNote)
-	}
-	if _, err := s.Store.Get(ctx, id); err != nil {
-		return nil, err
-	}
-	actor := httpauth.Actor(ctx)
-	if err := s.Store.RecordOutcome(ctx, outcome, actor, id, note); err != nil {
-		return nil, err
-	}
-	return s.Store.Usage(ctx, id)
-}
-
-// recordUsage writes usage events with the acting caller as provenance.
-// Failures are logged, never returned: usage recording must not fail reads.
-func (s *Service) recordUsage(ctx context.Context, event string, ids []string) {
-	if len(ids) == 0 {
-		return
-	}
-	if err := s.Store.RecordEvents(ctx, event, httpauth.Actor(ctx), ids); err != nil {
-		s.Log.Warn("usage recording failed", "event", event, "error", err)
-	}
-}
-
-// maxMissQuery bounds the query text kept for a search that found
-// nothing. A question a person would answer fits easily; past this the
-// string is a payload rather than a question, and the miss is worth
-// counting without keeping all of it (design doc 0051 §3.3).
-const maxMissQuery = 500
-
-// recordMiss keeps one unanswered question, with the acting caller as
-// provenance — same treatment as a usage event, and the same silence on
-// failure: nothing about measuring may fail a read.
-//
-// Deployments that keep no questions at all (public, or
-// OCHAKAI_RECORD_MISSES=false) stop here, not in the store: the decision
-// is about what this ochakai keeps, and the store is where things are
-// kept.
-func (s *Service) recordMiss(ctx context.Context, query string) {
-	if s.Config != nil && !s.Config.RecordMisses {
-		return
-	}
-	query = strings.TrimSpace(query)
-	if len(query) > maxMissQuery {
-		// On a rune boundary: a query is text somebody will read, and
-		// half a character is not.
-		query = strings.ToValidUTF8(query[:maxMissQuery], "")
-	}
-	if query == "" {
-		return
-	}
-	if err := s.Store.RecordMiss(ctx, query, httpauth.Actor(ctx)); err != nil {
-		s.Log.Warn("search miss recording failed", "error", err)
-	}
-}
-
-// maxStatsWindow is the longest window the stats will answer for, and it
-// is the raw-event retention (design doc 0029 §3.4): past it the events
-// and misses have been pruned, so a longer window would quietly answer
-// with less than it was asked for. defaultStatsWindow is a month —
-// long enough for a trend, short enough that a change shows in it.
-const (
-	maxStatsWindow     = 180
-	defaultStatsWindow = 30
-)
-
-// Stats returns the instance-level view of the improvement loop over the
-// last days days (design doc 0051 §3.5): what the base is made of, what
-// review did, what callers reported, and what they asked for and did not
-// find.
-//
-// days <= 0 means the default. A window past the retention is refused
-// rather than clamped: a caller asking for a year and receiving six
-// months' worth, silently, would read the answer as a year.
-//
-// prefixes scopes it to one or more subtrees (design doc 0041), which is
-// how a team on a shared deployment asks about its own knowledge — and
-// what let GET /api/v1/queues be folded into this face rather than
-// dropped (design doc 0049 §3.1). Only the misses ignore it: a question
-// that found nothing found it nowhere, so there is no id to scope by.
-func (s *Service) Stats(ctx context.Context, days int, prefixes []string) (*domain.Stats, error) {
-	if days == 0 {
-		days = defaultStatsWindow
-	}
-	if days < 0 || days > maxStatsWindow {
-		return nil, Invalidf("days must be between 1 and %d: raw events and misses are pruned after %d days, "+
-			"so nothing answers for the time before that", maxStatsWindow, maxStatsWindow)
-	}
-	f, err := checkedFilter(store.Filter{Prefixes: prefixes})
-	if err != nil {
-		return nil, err
-	}
-	now := time.Now().UTC()
-	st, err := s.Store.Stats(ctx, now.AddDate(0, 0, -days), f.Prefixes)
-	if err != nil {
-		return nil, err
-	}
-	st.At = now
-	st.WindowDays = days
-	st.Misses.Recording = s.Config == nil || s.Config.RecordMisses
-	return st, nil
-}
-
-// rrfFuse merges ranked lists with reciprocal rank fusion (k=60), adding a
-// small boost for verified concepts so certified knowledge surfaces first.
-func rrfFuse(limit int, lists ...[]domain.SearchHit) []domain.SearchHit {
-	const k = 60
-	type entry struct {
-		hit   domain.SearchHit
-		score float64
-	}
-	byKey := map[string]*entry{}
-	for _, list := range lists {
-		for rank, hit := range list {
-			key := hit.ID
-			e, ok := byKey[key]
-			if !ok {
-				e = &entry{hit: hit}
-				byKey[key] = e
-			}
-			e.score += 1.0 / float64(k+rank+1)
-		}
-	}
-	out := make([]domain.SearchHit, 0, len(byKey))
-	for _, e := range byKey {
-		// Any confirmation earns the nudge, whichever tier it is: the
-		// boost is about "somebody checked this", and weighting the tiers
-		// against each other would be a ranking decision no record makes
-		// (design docs 0033, 0046 §3.10).
-		if e.hit.Trust == domain.TrustHuman || e.hit.Trust == domain.TrustMachine {
-			e.score += 0.002
-		}
-		e.hit.Score = e.score
-		out = append(out, e.hit)
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].Score != out[j].Score {
-			return out[i].Score > out[j].Score
-		}
-		return out[i].ID < out[j].ID
-	})
-	if len(out) > limit {
-		out = out[:limit]
-	}
-	return out
-}
-
-// updateEmbedding refreshes the stored document vector. Failures are logged,
-// not returned: writes must not depend on the embedding provider being up.
-func (s *Service) updateEmbedding(ctx context.Context, k *domain.Knowledge) {
-	if s.Embedder == nil {
-		return
-	}
-	vecs, err := s.embedDocument(ctx, embeddingText(k, s.embedBytes()))
-	if err != nil {
-		s.Log.Warn("document embedding failed; concept remains findable by the lexical half of search", "type", k.Type, "id", k.ID, "error", err)
-		return
-	}
-	if err := s.Store.UpsertEmbedding(ctx, k.ID, s.Embedder.Model(), vecs[0]); err != nil {
-		s.Log.Warn("storing embedding failed", "id", k.ID, "error", err)
-	}
-}
-
-// embedBytes is how much text this deployment's model can be handed.
-//
-// The real limit is tokens, and UTF-8 bytes track tokens across scripts
-// far better than characters do — but only roughly, and the ratio is
-// worst where it matters most: Japanese runs three bytes per character
-// and can approach one token per character. So the cap is deliberately
-// conservative, covers the whole text (envelope included), and comes from
-// the model rather than from a constant: the window it has to fit inside
-// is a property of the model, and applying the smallest model's budget to
-// a model with four times the window throws away three quarters of it —
-// silently, since a concept truncated for no reason still embeds fine and
-// just carries less of itself.
-//
-// A provider that does not say (any embedder but Vertex, including the
-// fakes in tests) gets the conservative floor. The estimate is still an
-// estimate, so embedDocument retries shorter rather than trusting it.
-func (s *Service) embedBytes() int {
-	if l, ok := s.Embedder.(embed.InputLimited); ok {
-		return l.MaxInputBytes()
-	}
-	return embed.ConservativeInputBytes
-}
-
-// ReembedResult reports what a Reembed pass did. Missing is how many
-// concepts and attachments still have no vector once the pass is over —
-// the number that decides whether the operator runs it again, so it
-// counts what is actually left rather than what this pass did not get
-// to. Cursor is where the next pass resumes; empty means the corpus is
-// exhausted.
-type ReembedResult struct {
-	Embedded    int    `json:"embedded"`
-	Attachments int    `json:"attachments"`
-	Failed      int    `json:"failed"`
-	Missing     int    `json:"missing"`
-	Cursor      string `json:"cursor,omitempty"`
-}
-
-// A reembed pass is one HTTP request, and one concept is one call to the
-// embedding provider — gemini-embedding-001 takes a single text per
-// request, so the passes are sequential by construction. At a few hundred
-// milliseconds each, a pass of 500 is minutes of request time, and Cloud
-// Run's default request timeout is 300 seconds: a "bounded" pass that
-// cannot finish inside a request is not bounded, it just fails late and
-// reports nothing (the work it did is still written).
-//
-// So a pass is sized to finish, and the CLI repeats it until nothing is
-// left. The ceiling is for operators who know their own timeout.
-const (
-	defaultReembedPass = 200
-	maxReembedPass     = 5000
-)
-
-// Reembed fills in the vectors that concept writes never produced. Vectors
-// are written on create and update only, so a base loaded before semantic
-// search was reachable has none, and hybrid search silently stays
-// lexical-only until every concept happens to be rewritten. Switching models
-// has the same shape: the old vectors are in a space nothing queries any
-// more (design doc 0020), as does a changed dimension, which rebuilds the
-// tables empty (design doc 0053 §3).
-//
-// Failures are counted, not fatal: one concept the provider chokes on must
-// not strand the rest of the corpus.
-func (s *Service) Reembed(ctx context.Context, cursor string, limit int) (*ReembedResult, error) {
-	if err := s.readOnly(); err != nil {
-		return nil, err
-	}
-	if s.Embedder == nil {
-		return nil, Unsupportedf("semantic search is not enabled on this deployment: it is the default on Google Cloud once the service identity may call Vertex AI (design doc 0053)")
-	}
-	if limit <= 0 || limit > maxReembedPass {
-		limit = defaultReembedPass
-	}
-	entryCursor, attachCursor := splitReembedCursor(cursor)
-	ids, err := s.Store.ListUnembedded(ctx, s.Embedder.Model(), entryCursor, limit)
-	if err != nil {
-		return nil, err
-	}
-	res := &ReembedResult{}
-	for _, id := range ids {
-		k, err := s.Store.Get(ctx, id)
-		if err != nil {
-			res.Failed++
-			s.Log.Warn("reembed: concept disappeared", "id", id, "error", err)
-			continue
-		}
-		vecs, err := s.embedDocument(ctx, embeddingText(k, s.embedBytes()))
-		if err != nil {
-			res.Failed++
-			s.Log.Warn("reembed: embedding failed", "id", id, "error", err)
-			continue
-		}
-		if err := s.Store.UpsertEmbedding(ctx, id, s.Embedder.Model(), vecs[0]); err != nil {
-			res.Failed++
-			s.Log.Warn("reembed: storing failed", "id", id, "error", err)
-			continue
-		}
-		res.Embedded++
-	}
-	if len(ids) > 0 {
-		entryCursor = ids[len(ids)-1]
-	}
-	// Attachments carry vectors of their own (design doc 0020) and only
-	// attach writes them, so dropping the embedding tables to change
-	// dimensions leaves every file findable by name alone. Concepts come
-	// first and attachments fill the rest of the pass, so a corpus with
-	// both makes progress on both.
-	if rest := limit - len(ids); rest > 0 {
-		attachCursor, err = s.reembedAttachments(ctx, res, attachCursor, rest)
-		if err != nil {
-			return nil, err
-		}
-	}
-	// Counted after the pass, so it is what is left — not what was left
-	// minus what we did, which double-counts the work and reports a
-	// bounded pass as a finished one. Concepts this pass failed on are
-	// still unembedded and belong in the total.
-	missing, err := s.Store.CountUnembedded(ctx, s.Embedder.Model())
-	if err != nil {
-		// The pass itself succeeded; refusing to report it because the
-		// tally failed would be worse than reporting it without one.
-		s.Log.Warn("reembed: counting what is left failed", "error", err)
-		return res, nil
-	}
-	attMissing, err := s.Store.CountUnembeddedAttachments(ctx, s.Embedder.Model())
-	if err != nil {
-		s.Log.Warn("reembed: counting attachments left failed", "error", err)
-	}
-	res.Missing = missing + attMissing
-	if res.Missing > 0 {
-		res.Cursor = joinReembedCursor(entryCursor, attachCursor)
-	}
-	return res, nil
-}
-
-// reembedAttachments re-embeds up to limit attachments, returning the
-// cursor to resume from. Failures are counted like concept failures: a
-// file the provider rejects must not strand the rest.
-func (s *Service) reembedAttachments(ctx context.Context, res *ReembedResult, cursor string, limit int) (string, error) {
-	afterID, afterName := splitReembedCursor(cursor)
-	pending, err := s.Store.ListUnembeddedAttachments(ctx, s.Embedder.Model(), afterID, afterName, limit)
-	if err != nil {
-		return cursor, err
-	}
-	for _, a := range pending {
-		att, data, err := s.Store.GetAttachment(ctx, a.KnowledgeID, a.Name)
-		if err != nil {
-			res.Failed++
-			s.Log.Warn("reembed: attachment disappeared", "id", a.KnowledgeID, "name", a.Name, "error", err)
-			continue
-		}
-		// The same path attach uses, so the vectors match what a fresh
-		// attach would have produced; it logs and swallows its own
-		// failures, which is why the count below is of rows written.
-		before := s.attachmentVectorCount(ctx, a.KnowledgeID, a.Name)
-		s.updateAttachmentEmbedding(ctx, a.KnowledgeID, att, data)
-		if s.attachmentVectorCount(ctx, a.KnowledgeID, a.Name) > before {
-			res.Attachments++
-		} else {
-			res.Failed++
-		}
-		afterID, afterName = a.KnowledgeID, a.Name
-	}
-	return joinReembedCursor(afterID, afterName), nil
-}
-
-// attachmentVectorCount reports whether a current-model vector exists,
-// so a pass can tell an embedding that landed from one the provider (or
-// the model's file support) declined.
-func (s *Service) attachmentVectorCount(ctx context.Context, id, name string) int {
-	n, err := s.Store.CountAttachmentEmbedding(ctx, id, name, s.Embedder.Model())
-	if err != nil {
-		return 0
-	}
-	return n
-}
-
-// The reembed cursor carries two positions — concepts and attachments —
-// in one opaque string, so the wire stays a single token the CLI can
-// hand back unchanged. "\x00" cannot occur in an id (design doc 0017)
-// or an attachment name.
-func joinReembedCursor(a, b string) string {
-	if a == "" && b == "" {
-		return ""
-	}
-	return a + "\x00" + b
-}
-
-func splitReembedCursor(c string) (string, string) {
-	a, b, _ := strings.Cut(c, "\x00")
-	return a, b
-}
-
-// embedDocument embeds text, halving it and retrying while the model says
-// it is too long. Byte budgets only approximate a token count, and the
-// approximation is worst for Japanese; without this, a concept that
-// overruns is dropped from vector search silently — logged, but with the
-// write reported as a success. Halving converges in a couple of rounds
-// and only runs on the rejection, so an outage still costs one call.
-func (s *Service) embedDocument(ctx context.Context, text string) ([][]float32, error) {
-	// Four attempts halve a 5000-byte text to 625, well under any
-	// plausible window; the floor below stops it before it embeds a
-	// fragment too short to mean anything.
-	for attempt := range 4 {
-		vecs, err := s.Embedder.Embed(ctx, embed.TaskDocument, []string{text})
-		if err == nil {
-			if attempt > 0 {
-				s.Log.Info("embedded after shortening", "bytes", len(text), "attempts", attempt+1)
-			}
-			return vecs, nil
-		}
-		if !errors.Is(err, embed.ErrInputTooLong) || len(text) < 500 {
-			return nil, err
-		}
-		text = truncateUTF8(text, len(text)/2)
-	}
-	return nil, fmt.Errorf("still over the embedding model's input limit after shortening")
-}
-
-// embeddingText builds the document text to embed: envelope fields plus the
-// golden query question, body truncated to keep within model input limits.
-// The id leads (design doc 0022): with title optional, the filename may be
-// the concept's only name, and the path carries the domain hierarchy.
-func embeddingText(k *domain.Knowledge, max int) string {
-	parts := []string{k.ID, k.Title, k.Description, strings.Join(k.Tags, " ")}
-	if q, ok := k.Attrs["question"].(string); ok {
-		parts = append(parts, q)
-	}
-	parts = append(parts, k.Body)
-	// The cap covers the joined text: a concept whose envelope is long
-	// (a deep path, a paragraph of description) must not push the total
-	// past the window just because the body fit on its own.
-	return truncateUTF8(strings.TrimSpace(strings.Join(parts, "\n")), max)
-}
-
-// truncateUTF8 caps s at max bytes and then drops a trailing partial rune.
-// A plain s[:max] lands mid-character two times in three on Japanese text
-// (3 bytes per character), and half a character is not valid UTF-8.
-func truncateUTF8(s string, max int) string {
-	if len(s) > max {
-		s = s[:max]
-	}
-	for len(s) > 0 {
-		// DecodeLastRuneInString reports (RuneError, 1) for a byte that
-		// cannot end a valid encoding; a real U+FFFD decodes with size 3.
-		if r, size := utf8.DecodeLastRuneInString(s); r != utf8.RuneError || size > 1 {
-			break
-		}
-		s = s[:len(s)-1]
-	}
-	return s
-}
