+++ internal/store/store.go
@@ -10,14 +10,12 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"path"
 	"sort"
 	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
-	"unicode"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
@@ -149,59 +147,6 @@ func (s *Store) Close() {
 		s.flushWG.Wait()
 	}
 	s.pool.Close()
-}
-
-// Filter narrows search and list operations.
-type Filter struct {
-	Types    []domain.Type
-	Statuses []domain.Status
-	Tags     []string
-	// Source narrows to entries citing this exact resource — the reverse
-	// of sources[].resource, for "what derives from this material"
-	// (design doc 0037 §2.3). Single-valued on purpose: the question is
-	// about one artifact that changed, and a disjunction of containment
-	// tests would not use the GIN index.
-	Source string
-	// LinksTo narrows to entries whose body links at this id — the
-	// reverse of the edge design doc 0024 derives from the markdown,
-	// and the question the web UI asks as "linked from". Single-valued
-	// like Source and for the same reason: it is about one entry, and
-	// the answer is a set rather than a ranking.
-	LinksTo string
-	// Prefixes narrow to entries addressed under one of these paths —
-	// the id is the address (design doc 0017), so this is how a caller
-	// scopes a search to a team's subtree, the company-wide one, or both
-	// (design doc 0041). Repeatable and OR-ed, unlike Source: the ordinary
-	// question is "my scope and the shared one", and splitting that into
-	// two calls would leave the caller merging two incomparable rankings.
-	// Values arrive normalized by the service — NFC, no trailing slash.
-	Prefixes []string
-
-	// Trust and Rejected ask about the instance ledgers rather than the
-	// document (design docs 0043 §§3.2-3.3, 0046 §3.10). Trust is empty
-	// when the question is not asked, and OR-ed like Statuses when it is;
-	// Rejected is tri-state, and nil still hides rejected entries — the
-	// default is not "no constraint" but "not the ones we said no to".
-	Trust    []domain.Trust
-	Rejected *bool
-
-	// Frontmatter narrows by a frontmatter key read out of the jsonb
-	// index, which is what keeps the query surface additive: a key OKF
-	// adds in a later version is askable with no column and no migration
-	// (design doc 0046 §3.11). Which keys those are is the service's
-	// question, not the store's — OKF's own, less the ones a named filter
-	// answers (design doc 0047). Every pair must match (AND), matching a
-	// scalar exactly or a list by membership.
-	// Values are text, and a value that spells a number or a boolean
-	// matches the typed frontmatter as well as the text: YAML types what
-	// it parses, so `required: true` is indexed as a boolean and would
-	// otherwise be unaskable (frontmatterContainment).
-	//
-	// Exact and membership are the only two, deliberately: ranges,
-	// negation and boolean expressions are the doorway to a query
-	// language, and ochakai is a knowledge store whose answers are
-	// entries rather than a database whose answers are rows (0046 §5).
-	Frontmatter map[string]string
 }
 
 const knowledgeCols = `type, id, title, description, resource, tags, status, status_note, stale_after,
@@ -1082,1474 +1027,4 @@ func (s *Store) Purge(ctx context.Context, id string, actor domain.Actor) error 
 		}
 		return nil
 	})
-}
-
-// Move renames an entry to a new id. The id is the address (design doc
-// 0017), so everything keyed by it follows in one transaction — the row,
-// its revisions, attachments, usage, events, and embeddings — and live
-// entries that reference the old id (link targets in both bare and
-// both of SPEC §6's forms, and attrs.model) are rewritten so no reference
-// breaks. Attachment bytes never move: blobs are content-addressed
-// (design doc 0011). The destination must be a fresh id — a row there,
-// even soft-deleted, already owns that address and its revision history.
-func (s *Store) Move(ctx context.Context, oldID, newID string, actor domain.Actor) (*domain.Knowledge, error) {
-	k, err := s.Get(ctx, oldID)
-	if err != nil {
-		return nil, err
-	}
-	k.UpdatedAt = NowStored()
-	k.ContentChangedAt = k.UpdatedAt
-	// A move rewrites the entry's own relative links and every referrer's
-	// body, so the mover is who the content now stands by: generated.by
-	// in an export (design doc 0036 §3.3).
-	k.UpdatedBy = actor
-	err = s.withTx(ctx, func(tx pgx.Tx) error {
-		// A soft-deleted entry still owns the id — the row holds the primary
-		// key and its revisions hold (id, rev) — so the destination is taken
-		// either way. Create can revive such an entry in place; a move
-		// cannot, because the arriving entry brings revisions of its own.
-		// Say which case it is: "already exists" sends someone looking for
-		// an entry they cannot see, and without purge the id would be
-		// blocked forever.
-		var occupied, occupantDeleted bool
-		if err := tx.QueryRow(ctx,
-			`SELECT true, deleted_at IS NOT NULL FROM object WHERE id=$1`, newID).
-			Scan(&occupied, &occupantDeleted); err != nil && !errors.Is(err, pgx.ErrNoRows) {
-			return err
-		}
-		if occupied {
-			if occupantDeleted {
-				return fmt.Errorf("%w: %s holds a deleted entry; purge it to free the id", ErrAlreadyExists, newID)
-			}
-			return ErrAlreadyExists
-		}
-		// deleted_at IS NULL guards the race with a concurrent delete,
-		// exactly as in SoftDelete: the Get above ran outside this
-		// transaction.
-		tag, err := tx.Exec(ctx,
-			`UPDATE object SET id=$2, path=$8, updated_at=$3, content_changed_at=$3,
-			 updated_by_kind=$4, updated_by_name=$5, updated_by_via=$6, updated_by_producer=$7
-			 WHERE id=$1 AND deleted_at IS NULL`,
-			oldID, newID, k.UpdatedAt, actor.Kind, actor.Name, actor.Via, actor.Producer, domain.ConceptPath(newID))
-		if isUniqueViolation(err) {
-			// The probe above found the destination free, but it took no
-			// lock — a create can land on newID in the window. The primary
-			// key stops it either way; this is so the caller reads the
-			// same answer as when the id was already taken, rather than a
-			// 500 with a constraint name in it.
-			return fmt.Errorf("%w: %s", ErrAlreadyExists, newID)
-		}
-		if err != nil {
-			return err
-		}
-		if tag.RowsAffected() == 0 {
-			return ErrNotFound
-		}
-		for _, q := range []string{
-			// The entry's namespace moves with it (design doc 0046 §3.3):
-			// the address is part of what a file is, and leaving the
-			// files behind would leave them under a directory whose
-			// concept is gone. Only the files — a concept under there has
-			// an id, an address and a history of its own, and moves when
-			// somebody moves it.
-			//
-			// starts_with, not LIKE, for the reason browsing gives: an id
-			// may contain "_", which LIKE reads as a wildcard. Moving
-			// "sales_2024" would have dragged every file under
-			// "salesX2024/" along with it.
-			`UPDATE object SET path = $2 || substr(path, length($1) + 1)
-			 WHERE id IS NULL AND starts_with(path, $1 || '/')`,
-			// The naming follows the files it names.
-			`UPDATE object SET files = (
-			     SELECT COALESCE(jsonb_agg(
-			         CASE WHEN starts_with(x, $1 || '/') THEN $2 || substr(x, length($1) + 1) ELSE x END), '[]'::jsonb)
-			       FROM jsonb_array_elements_text(files) x)
-			 WHERE id IS NOT NULL
-			   AND EXISTS (SELECT 1 FROM jsonb_array_elements_text(files) x WHERE starts_with(x, $1 || '/'))`,
-			`UPDATE knowledge_revision SET id=$2, path=$2 || '.md' WHERE id=$1`,
-			// A file's history is keyed by its path, so it moves when the
-			// file does — otherwise the log of the directory it arrived
-			// in would be silent about how it got there, and the one it
-			// left would go on reporting a file that is not there.
-			`UPDATE knowledge_revision SET path = $2 || substr(path, length($1) + 1)
-			 WHERE id IS NULL AND starts_with(path, $1 || '/')`,
-			`UPDATE knowledge_verification SET id=$2 WHERE id=$1`,
-			`UPDATE knowledge_rejection SET id=$2 WHERE id=$1`,
-			`UPDATE attachment SET knowledge_id=$2 WHERE knowledge_id=$1`,
-			`UPDATE knowledge_usage SET knowledge_id=$2 WHERE knowledge_id=$1`,
-			`UPDATE knowledge_event SET knowledge_id=$2 WHERE knowledge_id=$1`,
-		} {
-			if _, err := tx.Exec(ctx, q, oldID, newID); err != nil {
-				return err
-			}
-		}
-		// Embedding tables exist only once semantic search has been
-		// enabled; the embedding text does not include the id, so
-		// re-keying is enough — no re-embed.
-		for _, q := range []string{
-			`UPDATE knowledge_embedding SET id=$2 WHERE id=$1`,
-			`UPDATE attachment_embedding SET knowledge_id=$2 WHERE knowledge_id=$1`,
-		} {
-			if err := execTolerateMissingTable(ctx, tx, q, oldID, newID); err != nil {
-				return err
-			}
-		}
-		k.ID = newID
-		if err := s.rewriteReferences(ctx, tx, oldID, k, actor); err != nil {
-			return err
-		}
-		return s.addRevision(ctx, tx, k, "move", actor)
-	})
-	if err != nil {
-		return nil, err
-	}
-	return k, nil
-}
-
-// rewriteReferences updates live entries that reference oldID — the
-// markdown links in their body, and attrs.model (design doc 0019) — each
-// rewrite recorded as an "update" revision. Runs after the rename itself,
-// so the moved entry (already renamed, passed as moved with the move's
-// timestamp stamped) is covered too — but its own rewrite is part of the
-// move, not a separate change: the row keeps the move's updated_at, no
-// "update" revision is added, and the rewritten body, links, and attrs
-// are folded back into moved so the caller's "move" revision and return
-// value carry the final state.
-//
-// The body is what gets rewritten: links are derived from it (design doc
-// 0024), so repairing the links column alone would leave the author's
-// prose pointing at an id that no longer exists. The links column is then
-// re-derived from the repaired body.
-//
-// Candidates come from the links column, which still holds the pre-move
-// derivation and so is an accurate index of who refers to oldID.
-func (s *Store) rewriteReferences(ctx context.Context, tx pgx.Tx, oldID string, moved *domain.Knowledge, actor domain.Actor) error {
-	newID := moved.ID
-	// FOR UPDATE, because each referrer is read here and written whole
-	// below: the rewrite carries body, links and attrs from this read, so
-	// an update committing in the window would be silently reverted — and
-	// recorded as an ordinary "update" revision, which hides that anything
-	// was lost. ORDER BY id gives concurrent moves one lock order, so they
-	// queue instead of deadlocking.
-	rows, err := tx.Query(ctx,
-		`SELECT `+knowledgeSelectDoc+` FROM object
-		 WHERE deleted_at IS NULL AND (links @> $1 OR attrs->>'model' = $2 OR id = $3)
-		 ORDER BY id FOR UPDATE`,
-		fmt.Sprintf(`[{"target": %q}]`, oldID),
-		oldID, newID)
-	if err != nil {
-		return err
-	}
-	// With the document: the rewrite below replaces the body inside the
-	// stored bytes, leaving the frontmatter of every entry that merely
-	// cited the moved one exactly as its writer left it.
-	referrers, err := pgx.CollectRows(rows, scanKnowledgeDoc)
-	if err != nil {
-		return err
-	}
-	now := NowStored()
-	movedDirs := path.Dir(oldID) != path.Dir(newID)
-	for i := range referrers {
-		r := &referrers[i]
-		self := r.ID == newID
-		// The moved entry resolves its own relative links against its old
-		// directory, so it is rewritten as if it still lived at oldID.
-		from := r.ID
-		body := r.Body
-		if self {
-			from = oldID
-			// Its own outbound relative links point at the old directory,
-			// and links are derived from the body against the entry's
-			// current id: left alone, "./gross.md" silently starts meaning
-			// a different entry (or none) the next time this row is
-			// written. Absolute is the form a move cannot reinterpret.
-			if movedDirs {
-				body = domain.AbsolutizeBodyLinks(oldID, body)
-			}
-		}
-		body = domain.RewriteBodyLinks(from, body, oldID, newID)
-		modelMoved := false
-		if m, ok := r.Attrs["model"].(string); ok && m == oldID {
-			r.Attrs["model"] = newID
-			modelMoved = true
-		}
-		if body == r.Body && !modelMoved {
-			continue // matched the links index but nothing to repair
-		}
-		r.Body = body
-		r.Doc = string(okf.ReplaceBody([]byte(r.Doc), body))
-		r.Links = domain.LinksFromBody(r.ID, r.Body)
-		// The rewrite is a content change by the mover, and it is recorded
-		// as one ("update" revision below), so the entry's generated.by
-		// follows it (design doc 0036 §3.3).
-		r.UpdatedBy = actor
-		// Only the columns a link rewrite touches are written: a move
-		// repairs references, and an entry's citations and computation
-		// contract are not references it makes.
-		j, err := marshalJSONFields(r)
-		if err != nil {
-			return err
-		}
-		// The moved entry's own rewrite is the move, not a change on top
-		// of it: keep the move's timestamp so the row, the "move"
-		// revision, and the returned entry agree on one instant.
-		if self {
-			r.UpdatedAt = moved.UpdatedAt
-		} else {
-			r.UpdatedAt = now
-		}
-		// A repaired link is a change to what the entry says, so
-		// generated moves with it — both halves of it, the actor above
-		// and the timestamp here (design doc 0046 §3.4).
-		r.ContentChangedAt = r.UpdatedAt
-		// deleted_at IS NULL restates what the locked read already
-		// established, so the statement does not depend on the reader
-		// having filtered: rewriting a tombstone would plant a repaired
-		// body and a revision on an entry nobody can see, to resurface
-		// whenever someone revives it.
-		// The stored document and its hash move with the body: the body
-		// is swapped inside the bytes the entry is stored as, and the
-		// index columns beside them are derived from the same value
-		// (design docs 0043 §3.1, 0046 §2.2).
-		doc, hash, err := storedDoc(r)
-		if err != nil {
-			return err
-		}
-		fm, err := storedFrontmatter(doc)
-		if err != nil {
-			return err
-		}
-		r.ContentHash = hash
-		tag, err := tx.Exec(ctx,
-			`UPDATE object SET links=$2, attrs=$3, body=$4, updated_at=$5,
-			 updated_by_kind=$6, updated_by_name=$7, updated_by_via=$8, updated_by_producer=$9,
-			 doc=$10, content_hash=$11, content_changed_at=$12, frontmatter=$13
-			 WHERE id=$1 AND deleted_at IS NULL`,
-			r.ID, j.links, j.attrs, r.Body, r.UpdatedAt, actor.Kind, actor.Name, actor.Via, actor.Producer, doc, hash,
-			r.ContentChangedAt, fm)
-		if err != nil {
-			return err
-		}
-		if tag.RowsAffected() == 0 {
-			return fmt.Errorf("rewriting references to %s: %s changed under the move", oldID, r.ID)
-		}
-		if self {
-			// Fold the result back into the caller's entry; the caller
-			// records the single "move" revision with this final state.
-			moved.Body = r.Body
-			moved.Links = r.Links
-			moved.Attrs = r.Attrs
-			// And the document the body was rewritten inside, with the
-			// version that goes with it: the "move" revision the caller
-			// records is written from this entry (design doc 0046 §2.2).
-			moved.Doc = r.Doc
-			moved.ContentHash = r.ContentHash
-			continue
-		}
-		if err := s.addRevision(ctx, tx, r, "update", actor); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// ListRevisions returns an entry's change history, newest first. It
-// reads history, so it works for soft-deleted entries too — the audit
-// trail is most interesting exactly when the entry is gone.
-func (s *Store) ListRevisions(ctx context.Context, id string, limit int) ([]domain.Revision, error) {
-	rows, err := s.pool.Query(ctx,
-		`SELECT rev, change, changed_by_kind, changed_by_name, changed_by_via, changed_by_producer, changed_at, doc
-		 FROM knowledge_revision WHERE id=$1 ORDER BY rev DESC LIMIT $2`,
-		id, limit)
-	if err != nil {
-		return nil, err
-	}
-	revs, err := pgx.CollectRows(rows, func(row pgx.CollectableRow) (domain.Revision, error) {
-		var r domain.Revision
-		return r, row.Scan(&r.Rev, &r.Change, &r.ChangedBy.Kind, &r.ChangedBy.Name, &r.ChangedBy.Via,
-			&r.ChangedBy.Producer, &r.ChangedAt, &r.Document)
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(revs) == 0 {
-		// Distinguish "no such entry" from "entry with empty history"
-		// (the latter cannot happen: create always writes rev 1).
-		return nil, ErrNotFound
-	}
-	return revs, nil
-}
-
-// LogRow is one line of a subtree's history: enough to render log.md
-// (SPEC §9) and no more. The document each revision holds is
-// deliberately not selected — a history that shipped every past version
-// of every entry in a directory would be the whole knowledge base
-// several times over, and what log.md publishes is what changed and by
-// whom (design doc 0046 §3.8).
-//
-// The row is keyed by the object's path, because that is what a
-// revision is an event about (design doc 0046 §3.1) and because a file
-// has no concept id — a log that keyed on the id would leave every
-// attach and detach out of the history it belongs in.
-type LogRow struct {
-	Path      string
-	Title     string
-	Change    string
-	ChangedBy domain.Actor
-	ChangedAt time.Time
-}
-
-// ListRevisionsAt returns the changes to the one object at this path,
-// newest first — the ledger behind ?history (design doc 0046 §3.5).
-//
-// Keyed by path rather than by concept id, so it answers for a file as
-// well: a revision is an event about an object (§3.1), and "what
-// happened to this object" should not be two questions depending on
-// what kind it is. It reads history, so a soft-deleted or purged-around
-// object still answers, which is when a history is most worth having.
-func (s *Store) ListRevisionsAt(ctx context.Context, path string, limit int) ([]LogRow, error) {
-	rows, err := s.pool.Query(ctx,
-		`SELECT r.path, COALESCE(k.title, ''), r.change,
-		        r.changed_by_kind, r.changed_by_name, r.changed_by_via, r.changed_by_producer, r.changed_at
-		 FROM knowledge_revision r
-		 LEFT JOIN object k ON k.path = r.path
-		 WHERE r.path = $1
-		 ORDER BY r.rev DESC LIMIT $2`, path, limit)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanLogRow)
-}
-
-// ListRevisionsUnder returns the newest changes anywhere under a path
-// prefix, newest first — concepts and files alike, which is the whole
-// point of one ledger keyed by path. An empty prefix is the whole
-// bundle.
-//
-// The title comes from the entry as it stands rather than from the
-// revision: a log names what changed, and naming it by a title it has
-// since stopped having would send a reader looking for something that is
-// not there. An entry that has since been purged keeps its history and
-// loses its title, which is why the join is left. A file has no title
-// and is named by its path.
-//
-// starts_with, not LIKE: a path may contain "_", which LIKE reads as a
-// wildcard (browse.go says the same about ids).
-func (s *Store) ListRevisionsUnder(ctx context.Context, prefix string, limit int) ([]LogRow, error) {
-	rows, err := s.pool.Query(ctx, revisionsUnderSQL, prefix, limit)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanLogRow)
-}
-
-// revisionsUnderSQL is the query behind both callers: the read at a
-// log.md address and the export writing the same file from a snapshot.
-// One statement, because two that drifted would be two histories.
-const revisionsUnderSQL = `SELECT r.path, COALESCE(k.title, ''), r.change,
-	        r.changed_by_kind, r.changed_by_name, r.changed_by_via, r.changed_by_producer, r.changed_at
-	 FROM knowledge_revision r
-	 LEFT JOIN object k ON k.path = r.path
-	 WHERE $1 = '' OR r.path = $1 || '.md' OR starts_with(r.path, $1 || '/')
-	 ORDER BY r.changed_at DESC, r.path, r.rev DESC LIMIT $2`
-
-func scanLogRow(row pgx.CollectableRow) (LogRow, error) {
-	var l LogRow
-	return l, row.Scan(&l.Path, &l.Title, &l.Change,
-		&l.ChangedBy.Kind, &l.ChangedBy.Name, &l.ChangedBy.Via, &l.ChangedBy.Producer, &l.ChangedAt)
-}
-
-func (s *Store) addRevision(ctx context.Context, tx pgx.Tx, k *domain.Knowledge, change string, actor domain.Actor) error {
-	// The full document, server-owned keys included: a revision records
-	// what the entry was, and who had confirmed it at the time is part of
-	// that (design doc 0043 §3.9).
-	doc, err := okf.Document(k)
-	if err != nil {
-		return err
-	}
-	// A revision is an event about an object, so the ledger counts by
-	// path (design doc 0046 §3.1): when a file's create and delete land
-	// here too, they share one history with the concept beside them.
-	_, err = tx.Exec(ctx, `INSERT INTO knowledge_revision
-		(path, id, rev, change, changed_by_kind, changed_by_name, changed_by_via, changed_by_producer, doc)
-		VALUES ($1, $2, (SELECT COALESCE(MAX(rev), 0) + 1 FROM knowledge_revision WHERE path=$1), $3, $4, $5, $6, $7, $8)`,
-		domain.ConceptPath(k.ID), k.ID, change, actor.Kind, actor.Name, actor.Via, actor.Producer, string(doc))
-	return err
-}
-
-// maxQueryFragments bounds how many pieces a query is split into: every
-// fragment is another ILIKE term on the same scan, and a pasted paragraph
-// would otherwise put a hundred of them there.
-const maxQueryFragments = 24
-
-// queryFragments splits a search query into the pieces a document might
-// actually contain.
-//
-// Matching the query as a whole is useless for the input get_context is
-// built for. Its documented argument is a data question, and the recall
-// hook feeds it a raw user prompt — "why is revenue down?" appears
-// verbatim in no body, and trigram similarity between a short query and a
-// multi-kilobyte document is 0.01-0.1 at best (for Japanese, where the
-// three-character windows of a question and a body almost never align, it
-// is routinely exactly 0). Neither a substring test nor a similarity
-// threshold can find anything from that. The signal is not the question;
-// it is the terms inside it.
-//
-// Latin-script tokens stay whole: "revenue" is a word, and its pieces are
-// not. Runs of Han/Hiragana/Katakana have no spaces to split on, so they
-// become sliding two-character windows — the standard n-gram treatment
-// for Japanese, and the smallest unit that still carries meaning (売上,
-// 受注, 原価).
-//
-// Two characters is below what a trigram index can serve, and pg_trgm
-// does not narrow the scan for such a pattern: LIKE '%売上%' has no word
-// boundary to pad against, so no whole trigram can be extracted and the
-// planner reads the table. Measured on 5000 entries, '%revenue%' and
-// '%売上の%' are index scans at 0.2ms while '%売上%' is a sequential scan
-// at 16ms. Three-character windows would restore the index and lose 売上,
-// 原価, 客数 — the terms the search is for — so the scan is the price of
-// finding them. SearchLexical keeps that scan cheap by scoring over ids
-// alone.
-func queryFragments(query string) []string {
-	// Fragments are collected per run and then interleaved, so the cap
-	// falls evenly across the query instead of truncating its tail. A
-	// question often names its subject last ("〜の件で、直近の原価率は?"),
-	// and cutting in reading order would drop exactly that.
-	var runs [][]string
-	for _, token := range strings.FieldsFunc(query, func(r rune) bool {
-		return !unicode.IsLetter(r) && !unicode.IsDigit(r)
-	}) {
-		// A token can still mix scripts, because FieldsFunc treats latin
-		// and kanji alike as letters: "PL科目", "売上KPI", "EC売上" all
-		// arrive whole. Dispatching on the first character alone would
-		// send "PL科目" through the latin path (matching only the exact
-		// string "PL科目") and cut "売上KPI" into windows that break KPI
-		// apart. Domain vocabulary is full of these, so split at the
-		// script boundary first and treat each run on its own terms.
-		for _, run := range scriptRuns(token) {
-			runes := []rune(run)
-			if !scriptWithoutSpaces(runes[0]) || len(runes) <= 2 {
-				runs = append(runs, []string{run})
-				continue
-			}
-			runs = append(runs, contentWindows(runes))
-		}
-	}
-
-	var out []string
-	seen := map[string]bool{}
-	for round := 0; len(out) < maxQueryFragments; round++ {
-		progressed := false
-		for _, run := range runs {
-			if round >= len(run) {
-				continue
-			}
-			progressed = true
-			if frag := run[round]; !seen[frag] {
-				seen[frag] = true
-				if out = append(out, frag); len(out) >= maxQueryFragments {
-					break
-				}
-			}
-		}
-		if !progressed {
-			break
-		}
-	}
-	if len(out) == 0 {
-		// A single character, or punctuation only: match it as written
-		// rather than matching nothing.
-		if q := strings.TrimSpace(query); q != "" {
-			out = append(out, q)
-		}
-	}
-	return out
-}
-
-// contentWindows cuts a space-less run into two-character windows and
-// keeps the ones that begin with a content character.
-//
-// Japanese grammar is written in hiragana and content words in kanji or
-// katakana, so a window's first character says which it is: 売上 and 下が
-// (the stem of 下がる) begin a word, while ぜ売 and が下 straddle a
-// boundary and mean nothing. Keeping only the former drops both the
-// function words, which match nearly every entry, and the accidental
-// fragments, which match almost none — and the latter matter more than
-// they look, because scoring weights rare fragments highest and a
-// meaningless one is rare by construction.
-//
-// A run with no content character at all — an all-kana query like
-// ください — keeps every window, or it would match nothing.
-func contentWindows(runes []rune) []string {
-	all := make([]string, 0, len(runes))
-	var content []string
-	for i := 0; i+2 <= len(runes); i++ {
-		w := string(runes[i : i+2])
-		all = append(all, w)
-		if contentChar(runes[i]) {
-			content = append(content, w)
-		}
-	}
-	if len(content) > 0 {
-		return content
-	}
-	return all
-}
-
-// scriptRuns splits a token wherever it crosses between a space-less
-// script (Han, kana) and anything else, so "PL科目" becomes "PL" and
-// "科目".
-func scriptRuns(token string) []string {
-	var runs []string
-	runes := []rune(token)
-	start := 0
-	for i := 1; i < len(runes); i++ {
-		if scriptWithoutSpaces(runes[i]) != scriptWithoutSpaces(runes[start]) {
-			runs = append(runs, string(runes[start:i]))
-			start = i
-		}
-	}
-	return append(runs, string(runes[start:]))
-}
-
-// scriptWithoutSpaces reports whether r belongs to a script that does not
-// separate words with spaces, so a token of it needs windowing rather than
-// taking whole.
-func scriptWithoutSpaces(r rune) bool {
-	return unicode.Is(unicode.Han, r) || unicode.Is(unicode.Hiragana, r) || unicode.Is(unicode.Katakana, r)
-}
-
-// contentChar reports whether r is the kind of character a Japanese
-// content word is written in, as opposed to the hiragana that spells the
-// grammar around it.
-func contentChar(r rune) bool {
-	return unicode.Is(unicode.Han, r) || unicode.Is(unicode.Katakana, r)
-}
-
-// SearchLexical ranks entries by how much of the query they contain,
-// verified entries boosted. The haystack is the search_text column
-// (migration 0016): the id (design doc 0022 — with title optional, the
-// filename may be the entry's only name), the envelope fields, the body,
-// and the attachment filenames (design doc 0020 — "seeds" finds the entry
-// carrying seeds.txt, embedder or not), maintained by trigger.
-//
-// The score is the fraction of the query's fragments the entry contains,
-// plus a bonus when the whole query appears verbatim (a keyword search
-// should outrank an entry that merely shares terms with a question), plus
-// the verified boost. An entry containing none of the fragments is not a
-// result — the floor is "matched something", not a magic number.
-//
-// Every term of that is a substring test against one column, so the
-// candidate predicate and the score read the same expressions, and the
-// GIN trigram index serves the ones it can — patterns of three characters
-// or more (migration 0016; two-character Japanese windows are answered by
-// a scan, see queryFragments). Fragment matching
-// replaced a similarity()/`%` pair that could not work: `%` is true only
-// above pg_trgm.similarity_threshold (0.3) and a real query never scores
-// that against a real document, so the candidate set was whatever the
-// whole-query substring test found — nothing, for a question.
-func (s *Store) SearchLexical(ctx context.Context, query string, f Filter, limit int) ([]domain.SearchHit, error) {
-	where, args := f.buildWhere("k.")
-	// Each pattern is escaped: unescaped, a query containing '%' matches
-	// every entry and flattens the ranking ('_' matches any one character).
-	args = append(args, "%"+escapeLike(query)+"%")
-	wholeParam := len(args)
-	frags := queryFragments(query)
-	var hit, weight, weighted, weights, tests []string
-	for i, frag := range frags {
-		args = append(args, "%"+escapeLike(frag)+"%")
-		tests = append(tests, fmt.Sprintf("k.search_text ILIKE $%d", len(args)))
-		hit = append(hit, fmt.Sprintf("k.search_text ILIKE $%d AS h%d", len(args), i))
-		// Document frequency over the candidates, not the whole table: one
-		// window pass, no extra scans.
-		weight = append(weight, fmt.Sprintf(
-			"ln(1 + total::float / GREATEST(count(*) FILTER (WHERE h%d) OVER (), 1)) AS w%d", i, i))
-		weighted = append(weighted, fmt.Sprintf("CASE WHEN h%d THEN w%d ELSE 0 END", i, i))
-		weights = append(weights, fmt.Sprintf("w%d", i))
-	}
-	// Scoring carries ids and booleans, never rows. The window functions
-	// force every candidate to be materialized before LIMIT can apply, and
-	// the candidate set is an OR over fragments — one common two-character
-	// window and it is most of the table. Selecting k.* through those
-	// layers would push every body, and search_text (a near-copy of the
-	// body) alongside it, through a sort node: tens of megabytes on a
-	// corpus of a few thousand, spilling work_mem on the instance size
-	// this is meant to run on. The entries are fetched once the top N is
-	// known.
-	q := fmt.Sprintf(`
-		WITH scored AS (
-			SELECT w.id, (%[1]s) / NULLIF(%[2]s, 0) + w.whole + w.verified AS score
-			FROM (
-				SELECT c.*, %[3]s FROM (
-					SELECT k.id, %[4]s, count(*) OVER () AS total,
-						CASE WHEN k.search_text ILIKE $%[5]d THEN 0.3 ELSE 0 END AS whole,
-						CASE WHEN EXISTS (SELECT 1 FROM knowledge_verification v
-							WHERE v.id = k.id) THEN 0.05 ELSE 0 END AS verified
-					FROM object k
-					WHERE (%[6]s) AND %[7]s
-				) c
-			) w
-			ORDER BY score DESC LIMIT %[8]d
-		)
-		SELECT `+knowledgeSelectK+`, scored.score
-		FROM object k JOIN scored ON scored.id = k.id
-		ORDER BY scored.score DESC`,
-		strings.Join(weighted, " + "), strings.Join(weights, " + "),
-		strings.Join(weight, ", "), strings.Join(hit, ", "), wholeParam,
-		strings.Join(tests, " OR "), where, limit)
-	rows, err := s.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanHit)
-}
-
-// SearchVector ranks by cosine distance against stored embeddings.
-func (s *Store) SearchVector(ctx context.Context, vec []float32, model string, f Filter, limit int) ([]domain.SearchHit, error) {
-	// Columns must be qualified: knowledge_embedding shares id/updated_at.
-	where, args := f.buildWhere("k.")
-	args = append(args, encodeVector(vec))
-	vecParam := len(args)
-	// Only this model's vectors: a vector written by another model lives
-	// in a different space, so comparing it to this query's vector yields
-	// a number that means nothing. Reembed refills the gap (design doc
-	// 0020); until it runs the entry is lexical-only, which is a visible
-	// absence rather than a plausible wrong ranking.
-	args = append(args, model)
-	q := fmt.Sprintf(`
-		SELECT `+knowledgeSelectK+`, 1 - (e.embedding <=> $%d::vector) AS score
-		FROM object k JOIN knowledge_embedding e ON k.id = e.id AND e.model = $%d
-		WHERE %s
-		ORDER BY e.embedding <=> $%d::vector LIMIT %d`, vecParam, len(args), where, vecParam, limit)
-	return s.annSearch(ctx, limit, q, args)
-}
-
-// SearchVectorAttachments ranks entries by their closest attachment
-// embedding (design doc 0020). Each entry appears once, carrying its
-// best attachment's score — attachments never stand alone, so the hit
-// is the owning entry.
-func (s *Store) SearchVectorAttachments(ctx context.Context, vec []float32, model string, f Filter, limit int) ([]domain.SearchHit, error) {
-	where, args := f.buildWhere("k.")
-	args = append(args, encodeVector(vec))
-	vecParam := len(args)
-	args = append(args, model) // this model's vectors only, as in SearchVector
-	q := fmt.Sprintf(`
-		SELECT `+withoutDoc(knowledgeCols)+", "+ledgerCols("best")+`, score FROM (
-			SELECT DISTINCT ON (k.id) k.*, 1 - (e.embedding <=> $%d::vector) AS score
-			FROM object k JOIN attachment_embedding e ON k.id = e.knowledge_id AND e.model = $%d
-			WHERE %s
-			ORDER BY k.id, e.embedding <=> $%d::vector
-		) best
-		ORDER BY score DESC LIMIT %d`, vecParam, len(args), where, vecParam, limit)
-	return s.annSearch(ctx, limit, q, args)
-}
-
-// efSearch is how many candidates the HNSW index is asked to consider for
-// a search wanting limit rows.
-//
-// The index scan runs first and the WHERE clause filters what it returns,
-// so ef_search is a ceiling on the answer, not a starting point: at
-// pgvector's default of 40, a search for 100 rows could never return more
-// than 40 even before a type or tag filter took its cut. Four times the
-// limit is the headroom for that filtering, floored at the default so a
-// small search is no worse than before and capped at pgvector's maximum.
-//
-// Set per query rather than on the connection: pooled connections are
-// shared, and a lingering session GUC would silently retune searches that
-// never asked.
-func efSearch(limit int) int {
-	const (
-		defaultEF = 40   // pgvector's own default
-		maxEF     = 1000 // pgvector's ceiling
-	)
-	ef := 4 * limit
-	if ef < defaultEF {
-		ef = defaultEF
-	}
-	if ef > maxEF {
-		ef = maxEF
-	}
-	return ef
-}
-
-// annSearch runs a vector query with ef_search sized for the limit. The
-// transaction exists only to scope SET LOCAL to this one statement; it
-// reads, so it ends in a rollback.
-func (s *Store) annSearch(ctx context.Context, limit int, q string, args []any) ([]domain.SearchHit, error) {
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = tx.Rollback(context.WithoutCancel(ctx)) }()
-	if _, err := tx.Exec(ctx, fmt.Sprintf("SET LOCAL hnsw.ef_search = %d", efSearch(limit))); err != nil {
-		return nil, err
-	}
-	rows, err := tx.Query(ctx, q, args...)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanHit)
-}
-
-// scanHit reads a knowledge row and projects it: a search result carries
-// the summary, not the entry (design doc 0043 §3.5). The row is still
-// selected whole — the projection is derived from the same columns every
-// other read uses, and narrowing the SELECT per caller would put a second
-// column list in the file to fall behind.
-func scanHit(row pgx.CollectableRow) (domain.SearchHit, error) {
-	var h domain.SearchHit
-	var k domain.Knowledge
-	dests, finish := knowledgeDest(&k)
-	if err := row.Scan(append(dests, &h.Score)...); err != nil {
-		return h, err
-	}
-	if err := finish(); err != nil {
-		return h, err
-	}
-	h.Summary = domain.SummaryOf(&k)
-	return h, nil
-}
-
-// buildWhere renders filter conditions; prefix qualifies columns (e.g. "k.")
-// for joined queries and may be empty. Rejected entries are excluded unless
-// asked for: knowledge that was never accepted must not resurface in
-// answers, but remains queryable on request so agents can check whether a
-// proposal was already rejected.
-func (f Filter) buildWhere(prefix string) (string, []any) {
-	// A file is an object in the same table now (design doc 0046 §3.13),
-	// and it is not an entry: no type, no title, nothing to rank. The id
-	// is what a concept has and a file has not, so every scan that
-	// returns entries says so here — once, where every filtered read goes
-	// through, rather than in each of them.
-	conds := []string{prefix + "deleted_at IS NULL", prefix + "id IS NOT NULL"}
-	var args []any
-	if len(f.Types) > 0 {
-		// Types match case-insensitively (design doc 0023 §3.3): the stored
-		// spelling is whatever the writer used, so both sides fold. There is
-		// no index on type, so lower() costs nothing here.
-		folded := make([]string, len(f.Types))
-		for i, t := range f.Types {
-			folded[i] = domain.FoldType(t)
-		}
-		args = append(args, folded)
-		conds = append(conds, fmt.Sprintf("lower(%stype) = ANY($%d)", prefix, len(args)))
-	}
-	if len(f.Statuses) > 0 {
-		args = append(args, f.Statuses)
-		conds = append(conds, fmt.Sprintf("%sstatus = ANY($%d)", prefix, len(args)))
-	}
-	// Rejection and verification are ledgers now, so both are asked about
-	// independently of the lifecycle value (design doc 0043 §§3.2-3.3).
-	// The rejected default is "hide", which is where it has always been —
-	// but it used to fall out of the status filter, and a filter that
-	// named any status turned the exclusion off with it. It no longer
-	// does: "show me the drafts" never meant "including the ones we
-	// turned down".
-	//
-	// The outer id must be qualified even when nothing else here is: the
-	// ledger tables have an id column of their own, so a bare "id" inside
-	// the subquery binds to the ledger's, making the condition r.id = r.id
-	// — true for every row, which silently empties every unaliased query.
-	outer := prefix
-	if outer == "" {
-		outer = "object."
-	}
-	rejectedExists := fmt.Sprintf(
-		"EXISTS (SELECT 1 FROM knowledge_rejection r WHERE r.id = %sid)", outer)
-	if f.Rejected != nil && *f.Rejected {
-		conds = append(conds, rejectedExists)
-	} else {
-		conds = append(conds, "NOT "+rejectedExists)
-	}
-	if len(f.Trust) > 0 {
-		// SPEC §5.3's tiers as predicates over the ledger: a person makes
-		// an entry human-reviewed, any other confirmation
-		// machine-confirmed, none unverified. Asking for several ORs them,
-		// so "anything somebody confirmed" is two values rather than a
-		// second kind of filter.
-		anyVerification := fmt.Sprintf(
-			"EXISTS (SELECT 1 FROM knowledge_verification v WHERE v.id = %sid)", outer)
-		byHuman := fmt.Sprintf(
-			"EXISTS (SELECT 1 FROM knowledge_verification v WHERE v.id = %sid AND v.by_kind = 'human')", outer)
-		var tiers []string
-		for _, t := range f.Trust {
-			switch t {
-			case domain.TrustUnverified:
-				tiers = append(tiers, "NOT "+anyVerification)
-			case domain.TrustMachine:
-				tiers = append(tiers, anyVerification+" AND NOT "+byHuman)
-			case domain.TrustHuman:
-				tiers = append(tiers, byHuman)
-			default:
-				// A tier nobody defines matches nothing. Dropping it
-				// instead would answer a question the caller did not ask
-				// — and answer it with everything.
-				tiers = append(tiers, "false")
-			}
-		}
-		if len(tiers) > 0 {
-			conds = append(conds, "("+strings.Join(tiers, ") OR (")+")")
-		}
-	}
-	if len(f.Tags) > 0 {
-		args = append(args, f.Tags)
-		conds = append(conds, fmt.Sprintf("%stags && $%d", prefix, len(args)))
-	}
-	// Containment against the frontmatter index, once per shape the value
-	// could have been written in (frontmatterContainment). Every shape is
-	// answered by the same GIN index, so the alternatives cost a wider
-	// index probe rather than a scan.
-	for _, key := range sortedKeys(f.Frontmatter) {
-		forms := frontmatterContainment(key, f.Frontmatter[key])
-		if len(forms) == 0 {
-			continue
-		}
-		alts := make([]string, 0, len(forms))
-		for _, form := range forms {
-			args = append(args, form)
-			alts = append(alts, fmt.Sprintf("%sfrontmatter @> $%d", prefix, len(args)))
-		}
-		conds = append(conds, "("+strings.Join(alts, " OR ")+")")
-	}
-	if f.Source != "" {
-		// Containment, so the GIN index on sources answers it directly.
-		// Matching is on resource alone: sources[].id is a key for the
-		// body's footnotes within one document, and means nothing across
-		// entries (design doc 0037 §2.3).
-		args = append(args, sourceContainment(f.Source))
-		conds = append(conds, fmt.Sprintf("%ssources @> $%d", prefix, len(args)))
-	}
-	if f.LinksTo != "" {
-		// The same containment the dedicated backlinks query used before
-		// this became a filter. There is no index on links, and there was
-		// none then either: the predicate is unchanged, and what changed
-		// is that it now composes with the rest of the filter instead of
-		// living in a query of its own.
-		args = append(args, linkContainment(f.LinksTo))
-		conds = append(conds, fmt.Sprintf("%slinks @> $%d", prefix, len(args)))
-	}
-	if cond, prefixArgs := prefixScope(prefix+"id", f.Prefixes, len(args)+1); cond != "" {
-		args = append(args, prefixArgs...)
-		conds = append(conds, cond)
-	}
-	return strings.Join(conds, " AND "), args
-}
-
-// prefixScope is the subtree condition on its own, over whichever column
-// holds the id. buildWhere is the whole of a listing's filter and carries
-// defaults a tally must not inherit — it hides rejected entries, and
-// stats counts them — so the one predicate the two do share is written
-// once here rather than twice, differently. Two faces that disagree about
-// what "teams/growth" covers is the failure this exists to prevent.
-//
-// Matching is on segment boundaries, not raw string prefixes: a path is a
-// sequence of segments, so "metrics" covers "metrics" itself and
-// everything under "metrics/" while leaving "metrics-legacy/churn" alone.
-// A raw prefix would pull a neighbouring directory into a scope the
-// caller drew around one (design doc 0041 §2.2). LIKE stays out for the
-// same reason browsing avoids it — ids may contain "_", which LIKE reads
-// as a wildcard.
-//
-// argN is the placeholder number the caller's next argument takes; the
-// empty prefix list is no condition at all.
-func prefixScope(column string, prefixes []string, argN int) (string, []any) {
-	if len(prefixes) == 0 {
-		return "", nil
-	}
-	return fmt.Sprintf(
-		`EXISTS (SELECT 1 FROM unnest($%d::text[]) AS p
-			WHERE %s = p OR left(%s, length(p) + 1) = p || '/')`,
-		argN, column, column), []any{prefixes}
-}
-
-// frontmatterContainment builds the jsonb documents that answer one
-// `fm.<key>=<value>` pair; the entry matches when it contains any of
-// them.
-//
-// A value the document wrote as a scalar and one it wrote inside a list
-// are both "the entry says k is v", and a caller asking `fm.tags=finance`
-// should not have to know which shape the writer chose.
-//
-// Both shapes are tried a second time when the value spells a JSON number
-// or boolean, because YAML types what it parses: `required: true` reaches
-// the index as the boolean true and `usage_count: 5` as the number 5,
-// while a filter value is always text. Text alone made every key whose
-// value is not a string unaskable — and unaskable silently, since a
-// containment test that matches nothing is zero rows rather than an
-// error, which is the wrong answer that looks like a valid one. Trying
-// both keeps §3.11's claim true for those keys: the day the spec adds
-// one, it can be asked for.
-//
-// It is not new syntax. `fm.n=5` is still one key and one value, it still
-// finds a document that wrote 5 as the string "5", and the comparison is
-// still exact match or list membership — the operators 0046 §5 keeps out
-// stay out.
-func frontmatterContainment(key, value string) [][]byte {
-	forms := []any{value, []string{value}}
-	if typed, ok := jsonScalar(value); ok {
-		forms = append(forms, typed, []json.RawMessage{typed})
-	}
-	out := make([][]byte, 0, len(forms))
-	for _, form := range forms {
-		b, err := json.Marshal(map[string]any{key: form})
-		if err != nil { // a string, a number and a boolean all marshal
-			continue
-		}
-		out = append(out, b)
-	}
-	return out
-}
-
-// jsonScalar returns value's typed JSON spelling and reports whether it
-// has one. A number and the two booleans do, and nothing else does:
-// those are the YAML scalars that reach the frontmatter index as
-// something other than text (okf.yamlScalar renders a date back as a
-// string, so a date needs no second form).
-//
-// An object and an array are deliberately not read. A filter value is a
-// value, and one that could carry a structure would be the query language
-// 0046 §5 closed the door on, arriving through the value instead of
-// through an operator.
-//
-// The text is passed through verbatim rather than decoded and re-encoded,
-// so an integer longer than a float64 can hold keeps its digits.
-func jsonScalar(value string) (json.RawMessage, bool) {
-	switch value {
-	case "true", "false":
-		return json.RawMessage(value), true
-	}
-	// A leading digit or minus is what separates a number from the rest of
-	// the grammar: json.Valid also accepts `"5"`, `[5]` and `null`, none of
-	// which is a scalar a producer wrote unquoted.
-	if value == "" || !json.Valid([]byte(value)) {
-		return nil, false
-	}
-	if c := value[0]; c != '-' && (c < '0' || c > '9') {
-		return nil, false
-	}
-	return json.RawMessage(value), true
-}
-
-// sourceContainment builds the jsonb value that matches an entry citing
-// resource. Marshalled rather than formatted, so a resource containing a
-// quote or a backslash cannot break out of the literal.
-func sourceContainment(resource string) []byte {
-	b, err := json.Marshal([]map[string]string{{"resource": resource}})
-	if err != nil { // a string always marshals
-		panic(fmt.Sprintf("marshalling source filter %q: %v", resource, err))
-	}
-	return b
-}
-
-// linkContainment builds the jsonb document that matches an entry whose
-// links name the given target — the containment the reverse lookup asks
-// with.
-func linkContainment(target string) []byte {
-	b, err := json.Marshal([]map[string]string{{"target": target}})
-	if err != nil { // a string always marshals
-		panic(fmt.Sprintf("marshalling links filter %q: %v", target, err))
-	}
-	return b
-}
-
-// After is the position a listing resumes from: the values of the
-// ordering columns for the last row handed out, in the order they are
-// ordered by, with the id last (design doc 0050 §2.1). A nil element is
-// SQL NULL — the never-verified entries at the tail of an ASC NULLS LAST
-// order. Nil After is the first page.
-//
-// The values travel as text and are cast back per column, which is what
-// lets the same shape carry a timestamp, a date and a count.
-type After struct {
-	Keys []*string
-	ID   string
-}
-
-// orderCol describes one column of a listing's ORDER BY, so the same
-// keyset predicate can be built for every feed rather than once per
-// feed. cast is the type the cursor's text is compared as.
-type orderCol struct {
-	expr string
-	cast string
-	desc bool
-	// null marks a column ordered ASC NULLS LAST — the one nullable
-	// position any of the feeds has (verification age).
-	null bool
-}
-
-// keysetAfter renders "strictly after this position" for an ORDER BY of
-// cols followed by idExpr, appending the cursor's values to args. It
-// returns "" when there is no position to resume from, so the caller can
-// AND it in unconditionally.
-//
-// The predicate is the lexicographic comparison spelled out rather than
-// PostgreSQL's row constructor: a row comparison cannot express one
-// column descending beside another ascending, and every feed here mixes
-// the two (most-failed first, fewest corroborations first).
-func keysetAfter(cols []orderCol, idExpr string, a *After, args *[]any) (string, error) {
-	if a == nil {
-		return "", nil
-	}
-	if len(a.Keys) != len(cols) {
-		return "", fmt.Errorf("cursor carries %d keys, this listing orders by %d", len(a.Keys), len(cols))
-	}
-	// Built from the id outwards: the innermost term is "same on every
-	// key, later id", and each column wraps it with "strictly after on
-	// this key, or equal on it and after by everything below".
-	*args = append(*args, a.ID)
-	pred := fmt.Sprintf("%s > $%d", idExpr, len(*args))
-	for i := len(cols) - 1; i >= 0; i-- {
-		c, v := cols[i], a.Keys[i]
-		var strict, eq string
-		if v == nil {
-			// Nothing sorts after a NULL in ASC NULLS LAST, so the walk
-			// can only continue among the other NULLs, by id.
-			if !c.null {
-				return "", fmt.Errorf("cursor holds no value for %s, which is never null", c.expr)
-			}
-			strict, eq = "false", c.expr+" IS NULL"
-		} else {
-			*args = append(*args, *v)
-			ref := fmt.Sprintf("$%d::%s", len(*args), c.cast)
-			op := ">"
-			if c.desc {
-				op = "<"
-			}
-			strict = fmt.Sprintf("%s %s %s", c.expr, op, ref)
-			if c.null {
-				// A NULL sorts after every value, so it is after this one.
-				strict = fmt.Sprintf("(%s OR %s IS NULL)", strict, c.expr)
-			}
-			eq = fmt.Sprintf("%s = %s", c.expr, ref)
-		}
-		pred = fmt.Sprintf("(%s OR (%s AND %s))", strict, eq, pred)
-	}
-	return " AND " + pred, nil
-}
-
-// ListByAddress lists the entries a filter matches, by id. It is the
-// listing the reverse lookups take: "what cites this resource" (design
-// doc 0037 §2.3) and "what links at this entry" are both questions with
-// no text to rank by, so they list rather than search. Ordered by id
-// because the answer is a set — a stable, readable order beats a
-// relevance score nothing computed, and it is a total order, so a cursor
-// can resume it (design doc 0050).
-//
-// Named for the order and not for either filter: which entries it
-// returns is the filter's business, and it had a filter's name while
-// serving two.
-func (s *Store) ListByAddress(ctx context.Context, f Filter, after *After, limit int) ([]domain.Knowledge, error) {
-	where, args := f.buildWhere("")
-	keyset, err := keysetAfter(nil, "id", after, &args)
-	if err != nil {
-		return nil, err
-	}
-	return s.queryKnowledge(ctx, fmt.Sprintf(`SELECT `+knowledgeSelect+` FROM object WHERE %s%s
-		ORDER BY id LIMIT %d`, where, keyset, limit), args...)
-}
-
-// ListByStaleAfter returns the entries whose declared expiry has passed,
-// most overdue first — the feed for "what did we say needed re-checking
-// by now" (design doc 0037 §2.1).
-//
-// Only entries that are stale today. A declaration that has not come due
-// is not work, and 0025 asked feeds to be queues a reviewer can finish;
-// the same reason sort=failed lists only entries with failures.
-//
-// Unlike the other two feeds, verifying does not empty this one: the date
-// is the writer's declaration, not something the server observed, so
-// clearing it means editing the entry to re-declare an expiry
-// (design doc 0037 §2.2).
-func (s *Store) ListByStaleAfter(ctx context.Context, f Filter, after *After, limit int) ([]domain.Knowledge, error) {
-	where, args := f.buildWhere("")
-	keyset, err := keysetAfter(
-		[]orderCol{{expr: "stale_after", cast: "date"}}, "id", after, &args)
-	if err != nil {
-		return nil, err
-	}
-	return s.queryKnowledge(ctx, fmt.Sprintf(`SELECT `+knowledgeSelect+` FROM object WHERE %s
-		AND `+pastExpiry("")+`%s
-		ORDER BY stale_after ASC, id LIMIT %d`, where, keyset, limit), args...)
-}
-
-// ListByVerifiedAt returns filtered entries ordered by verification age,
-// oldest first (never-verified entries last). This is the feed for golden
-// query canary runs: "which verified queries have gone longest unchecked".
-func (s *Store) ListByVerifiedAt(ctx context.Context, f Filter, after *After, limit int) ([]domain.Knowledge, error) {
-	where, args := f.buildWhere("")
-	keyset, err := keysetAfter(
-		[]orderCol{{expr: lastVerifiedAt("object"), cast: "timestamptz", null: true}},
-		"id", after, &args)
-	if err != nil {
-		return nil, err
-	}
-	return s.queryKnowledge(ctx, fmt.Sprintf(`SELECT `+knowledgeSelect+` FROM object WHERE %s%s
-		ORDER BY `+lastVerifiedAt("object")+` ASC NULLS LAST, id LIMIT %d`, where, keyset, limit), args...)
-}
-
-// usageLateral aggregates a knowledge entry's per-event running totals
-// (see usage.go) in one pass, exposing them as u.search_hits … u.failed and
-// u.last_used_at. Shared by the usage-carrying list feeds (ListByUsage,
-// ListByFailed), which select the same projection but order it differently.
-const usageLateral = `
-	LEFT JOIN LATERAL (
-		SELECT
-			COALESCE(sum(count) FILTER (WHERE event = 'search_hit'), 0) AS search_hits,
-			COALESCE(sum(count) FILTER (WHERE event = 'fetched'), 0)   AS fetches,
-			COALESCE(sum(count) FILTER (WHERE event = 'worked'), 0)    AS worked,
-			COALESCE(sum(count) FILTER (WHERE event = 'failed'), 0)    AS failed,
-			max(last_at) FILTER (WHERE event = 'failed')                AS last_failed_at,
-			max(last_at) AS last_used_at
-		FROM knowledge_usage
-		WHERE knowledge_id = k.id
-	) u ON true`
-
-// ListByUsage returns filtered entries ordered by demand, most-searched
-// first: search_hits descending, then oldest-created (created_at ascending)
-// as the tiebreak. This is the draft review feed — the promotion queue at
-// the top, and never-used drafts (search_hits 0) sinking oldest-first to
-// the bottom for inventory. Each hit carries its usage totals so the
-// caller renders the signal without a per-entry round trip. Score is 0.
-func (s *Store) ListByUsage(ctx context.Context, f Filter, after *After, limit int) ([]domain.SearchHit, error) {
-	where, args := f.buildWhere("k.")
-	keyset, err := keysetAfter([]orderCol{
-		{expr: "u.search_hits", cast: "bigint", desc: true},
-		{expr: "k.created_at", cast: "timestamptz"},
-	}, "k.id", after, &args)
-	if err != nil {
-		return nil, err
-	}
-	q := fmt.Sprintf(`
-		SELECT `+knowledgeSelectK+`,
-			u.search_hits, u.fetches, u.worked, u.failed, u.last_used_at
-		FROM object k`+usageLateral+`
-		WHERE %s%s
-		ORDER BY u.search_hits DESC, k.created_at ASC, k.id LIMIT %d`, where, keyset, limit)
-	rows, err := s.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanUsageHit)
-}
-
-// ListByFailed returns filtered entries whose failure reports are still
-// unanswered, worst first — the re-verification feed (design doc 0025).
-// It is the evidence-based counterpart to the verified_at feed's
-// time-based staleness, so a healthy base yields an empty feed.
-//
-// "Still unanswered" is the difference between a queue and a ledger. The
-// totals in knowledge_usage only ever grow, so ranking by them alone left
-// an entry in the feed forever: a reviewer who checked it, found it right
-// and re-verified it saw it sitting at the top the next morning, and the
-// queue could only lengthen. An entry drops out once it is verified after
-// its last failure report (Verify, or a promotion from draft) — or once
-// it is rejected, which the default status filter already hides.
-// Never-verified entries stay: a failing draft has had no ruling at all.
-//
-// Ordering stays on the lifetime failure count: presence answers "is this
-// unresolved", ranking answers "how bad has this been". Ties break on
-// fewest corroborating "worked" reports, then verification age (oldest
-// first, never-verified last), then id. Each hit carries its usage totals
-// so the reviewer sees the worked/failed evidence inline; score is 0.
-func (s *Store) ListByFailed(ctx context.Context, f Filter, after *After, limit int) ([]domain.SearchHit, error) {
-	where, args := f.buildWhere("k.")
-	keyset, err := keysetAfter([]orderCol{
-		{expr: "u.failed", cast: "bigint", desc: true},
-		{expr: "u.worked", cast: "bigint"},
-		{expr: lastVerifiedAt("k"), cast: "timestamptz", null: true},
-	}, "k.id", after, &args)
-	if err != nil {
-		return nil, err
-	}
-	q := fmt.Sprintf(`
-		SELECT `+knowledgeSelectK+`,
-			u.search_hits, u.fetches, u.worked, u.failed, u.last_used_at
-		FROM object k`+usageLateral+`
-		WHERE %s AND `+unansweredFailure("k")+`%[4]s
-		ORDER BY u.failed DESC, u.worked ASC, %[2]s ASC NULLS LAST, k.id LIMIT %[3]d`,
-		where, lastVerifiedAt("k"), limit, keyset)
-	rows, err := s.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, scanUsageHit)
-}
-
-// QueueCounts counts the three review queues in one pass: how many
-// entries each of ListByUsage's draft feed, ListByFailed and
-// ListByStaleAfter would return (design doc 0049). A count and a listing
-// want different queries — measuring three feeds by fetching a thousand
-// rows from each is not measuring — but they must not want different
-// *predicates*: those are the feeds' own (unansweredFailure, pastExpiry),
-// so a queue's depth cannot come to mean something other than the queue.
-//
-// The filter is the same one the feeds take, so a count is scoped by
-// prefix the way the listings are: a team asks about its own subtree.
-func (s *Store) QueueCounts(ctx context.Context, f Filter) (domain.QueueCounts, error) {
-	where, args := f.buildWhere("k.")
-	q := fmt.Sprintf(`
-		SELECT
-			count(*) FILTER (WHERE k.status = 'draft'),
-			count(*) FILTER (WHERE `+unansweredFailure("k")+`),
-			count(*) FILTER (WHERE `+pastExpiry("k.")+`)
-		FROM object k`+usageLateral+`
-		WHERE %[1]s`, where)
-	var c domain.QueueCounts
-	err := s.pool.QueryRow(ctx, q, args...).
-		Scan(&c.Drafts, &c.Failed, &c.StaleAfter)
-	return c, err
-}
-
-// scanUsageHit reads a knowledge row followed by its usage totals (the
-// ListByUsage projection). Score stays 0 — this is a listing, not a search.
-func scanUsageHit(row pgx.CollectableRow) (domain.SearchHit, error) {
-	var h domain.SearchHit
-	var k domain.Knowledge
-	var u domain.Usage
-	dests, finish := knowledgeDest(&k)
-	if err := row.Scan(append(dests,
-		&u.SearchHits, &u.Fetches, &u.Worked, &u.Failed, &u.LastUsedAt)...); err != nil {
-		return h, err
-	}
-	if err := finish(); err != nil {
-		return h, err
-	}
-	h.Summary = domain.SummaryOf(&k)
-	h.Usage = &u
-	return h, nil
-}
-
-// ListUnembedded returns the ids of live entries with no vector for the
-// named model, oldest first, up to limit. That covers both halves of the
-// same gap: entries written before semantic search was configured (there
-// is no row), and entries whose vector belongs to a model that has since
-// been changed (design doc 0020 — the old vectors sit in a space nobody
-// queries any more).
-//
-// after is a keyset cursor: pass the last id of the previous page to get
-// the next one. Without it a pass whose entries all fail keeps handing
-// back the same window, and the caller — which stops when a pass embeds
-// nothing — never reaches the rest of the corpus.
-func (s *Store) ListUnembedded(ctx context.Context, model, after string, limit int) ([]string, error) {
-	rows, err := s.pool.Query(ctx,
-		`SELECT k.id FROM object k
-		 LEFT JOIN knowledge_embedding e ON e.id = k.id AND e.model = $1
-		 WHERE k.deleted_at IS NULL AND k.id IS NOT NULL AND e.id IS NULL AND ($2 = '' OR k.id > $2)
-		 ORDER BY k.id LIMIT $3`, model, after, limit)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, pgx.RowTo[string])
-}
-
-// UnembeddedAttachment names one attachment awaiting a vector.
-type UnembeddedAttachment struct {
-	KnowledgeID string
-	Name        string
-}
-
-// ListUnembeddedAttachments is ListUnembedded for attachment vectors
-// (design doc 0020). Attach writes them, so a corpus loaded before
-// semantic search — or one whose embedding tables were dropped to change
-// dimensions — has none, and the files are findable by name only. The
-// cursor is the (knowledge_id, name) pair, for the same reason.
-func (s *Store) ListUnembeddedAttachments(ctx context.Context, model, afterID, afterName string, limit int) ([]UnembeddedAttachment, error) {
-	rows, err := s.pool.Query(ctx,
-		`SELECT k.id, split_part(f.path, '/', array_length(string_to_array(f.path, '/'), 1))
-		   FROM object k JOIN object f ON `+attributedTo+`
-		   LEFT JOIN attachment_embedding e
-		     ON e.knowledge_id = k.id
-		    AND e.name = split_part(f.path, '/', array_length(string_to_array(f.path, '/'), 1))
-		    AND e.model = $1
-		  WHERE k.id IS NOT NULL AND k.deleted_at IS NULL AND e.knowledge_id IS NULL
-		    AND ($2 = '' OR (k.id, f.path) > ($2, $3))
-		  ORDER BY k.id, f.path LIMIT $4`, model, afterID, afterName, limit)
-	if err != nil {
-		if isUndefinedTable(err) {
-			return nil, nil // embedding tables appear with semantic search
-		}
-		return nil, err
-	}
-	out, err := pgx.CollectRows(rows, func(row pgx.CollectableRow) (UnembeddedAttachment, error) {
-		var a UnembeddedAttachment
-		return a, row.Scan(&a.KnowledgeID, &a.Name)
-	})
-	if err != nil && isUndefinedTable(err) {
-		return nil, nil
-	}
-	return out, err
-}
-
-// CountUnembeddedAttachments counts what ListUnembeddedAttachments would
-// eventually return.
-func (s *Store) CountUnembeddedAttachments(ctx context.Context, model string) (int, error) {
-	var n int
-	err := s.pool.QueryRow(ctx,
-		`SELECT count(*)
-		   FROM object k JOIN object f ON `+attributedTo+`
-		   LEFT JOIN attachment_embedding e
-		     ON e.knowledge_id = k.id
-		    AND e.name = split_part(f.path, '/', array_length(string_to_array(f.path, '/'), 1))
-		    AND e.model = $1
-		  WHERE k.id IS NOT NULL AND k.deleted_at IS NULL AND e.knowledge_id IS NULL`, model).Scan(&n)
-	if err != nil && isUndefinedTable(err) {
-		return 0, nil
-	}
-	return n, err
-}
-
-// CountUnembedded counts the live entries with no vector for the named
-// model. Reembed reports it so a bounded pass cannot be mistaken for a
-// finished one.
-func (s *Store) CountUnembedded(ctx context.Context, model string) (int, error) {
-	var n int
-	err := s.pool.QueryRow(ctx,
-		`SELECT count(*) FROM object k
-		 LEFT JOIN knowledge_embedding e ON e.id = k.id AND e.model = $1
-		 WHERE k.deleted_at IS NULL AND k.id IS NOT NULL AND e.id IS NULL`, model).Scan(&n)
-	return n, err
-}
-
-// UpsertEmbedding stores the document embedding for a knowledge entry.
-func (s *Store) UpsertEmbedding(ctx context.Context, id, model string, vec []float32) error {
-	_, err := s.pool.Exec(ctx, `INSERT INTO knowledge_embedding (id, model, embedding, updated_at)
-		VALUES ($1, $2, $3::vector, now())
-		ON CONFLICT (id) DO UPDATE SET model = $2, embedding = $3::vector, updated_at = now()`,
-		id, model, encodeVector(vec))
-	return err
-}
-
-// UpsertAttachmentEmbedding stores the document embedding for one
-// attachment (design doc 0020).
-func (s *Store) UpsertAttachmentEmbedding(ctx context.Context, id, name, model string, vec []float32) error {
-	_, err := s.pool.Exec(ctx, `INSERT INTO attachment_embedding (knowledge_id, name, model, embedding, updated_at)
-		VALUES ($1, $2, $3, $4::vector, now())
-		ON CONFLICT (knowledge_id, name) DO UPDATE SET model = $3, embedding = $4::vector, updated_at = now()`,
-		id, name, model, encodeVector(vec))
-	return err
-}
-
-// execTolerateMissingTable runs sql inside tx, treating an undefined
-// table as a no-op. Embedding tables only exist once semantic search has
-// been enabled, and a failed statement aborts the whole Postgres
-// transaction, so the statement runs under a savepoint.
-func execTolerateMissingTable(ctx context.Context, tx pgx.Tx, sql string, args ...any) error {
-	sp, err := tx.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	if _, err := sp.Exec(ctx, sql, args...); err != nil {
-		_ = sp.Rollback(ctx)
-		if !isUndefinedTable(err) {
-			return err
-		}
-		return nil
-	}
-	return sp.Commit(ctx)
-}
-
-func (s *Store) withTx(ctx context.Context, fn func(tx pgx.Tx) error) error {
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	if err := fn(tx); err != nil {
-		_ = tx.Rollback(ctx)
-		return err
-	}
-	return tx.Commit(ctx)
-}
-
-// knowledgeJSON carries the entry's jsonb columns, encoded once per write.
-// The three nullable ones stay nil when the entry declares nothing, so an
-// undeclared usage_window or contract is SQL NULL rather than the JSON
-// literal "null" — the read side tells those apart to keep "no window" and
-// "an empty window" distinct (design doc 0036 §3.1).
-type knowledgeJSON struct {
-	links, attrs                    []byte
-	sources, parameters             []byte
-	usageWindow, executor, attester []byte
-}
-
-func marshalJSONFields(k *domain.Knowledge) (knowledgeJSON, error) {
-	if k.Tags == nil {
-		k.Tags = []string{}
-	}
-	if k.Links == nil {
-		k.Links = []domain.Link{}
-	}
-	if k.Attrs == nil {
-		k.Attrs = map[string]any{}
-	}
-	if k.Sources == nil {
-		k.Sources = []domain.Source{}
-	}
-	if k.Parameters == nil {
-		k.Parameters = []domain.Parameter{}
-	}
-	var j knowledgeJSON
-	for _, f := range []struct {
-		out *[]byte
-		in  any
-		// nullable columns encode only when the entry has the value.
-		skip bool
-	}{
-		{&j.links, k.Links, false},
-		{&j.attrs, k.Attrs, false},
-		{&j.sources, k.Sources, false},
-		{&j.parameters, k.Parameters, false},
-		{&j.usageWindow, k.UsageWindow, k.UsageWindow == nil},
-		{&j.executor, k.Executor, k.Executor == nil},
-		{&j.attester, k.Attester, k.Attester == nil},
-	} {
-		if f.skip {
-			continue
-		}
-		b, err := json.Marshal(f.in)
-		if err != nil {
-			return knowledgeJSON{}, err
-		}
-		*f.out = b
-	}
-	return j, nil
-}
-
-// escapeLike neutralizes the LIKE/ILIKE pattern metacharacters '%' and '_'
-// in a user query so it matches literally. PostgreSQL's default LIKE escape
-// character is the backslash, so the backslash itself is escaped first.
-func escapeLike(s string) string {
-	return strings.NewReplacer(`\`, `\\`, `%`, `\%`, `_`, `\_`).Replace(s)
-}
-
-// encodeVector renders a pgvector literal like "[0.1,0.2]".
-func encodeVector(vec []float32) string {
-	var b strings.Builder
-	b.WriteByte('[')
-	for i, v := range vec {
-		if i > 0 {
-			b.WriteByte(',')
-		}
-		fmt.Fprintf(&b, "%g", v)
-	}
-	b.WriteByte(']')
-	return b.String()
-}
-
-func isUndefinedTable(err error) bool {
-	return err != nil && strings.Contains(err.Error(), "42P01")
-}
-
-// CountAttachmentEmbedding reports whether one attachment has a vector
-// for the named model (0 or 1). Reembed uses it to tell an embedding
-// that landed from one the provider declined — the attach path logs its
-// own failures rather than returning them.
-func (s *Store) CountAttachmentEmbedding(ctx context.Context, id, name, model string) (int, error) {
-	var n int
-	err := s.pool.QueryRow(ctx,
-		`SELECT count(*) FROM attachment_embedding
-		 WHERE knowledge_id = $1 AND name = $2 AND model = $3`, id, name, model).Scan(&n)
-	if err != nil && isUndefinedTable(err) {
-		return 0, nil
-	}
-	return n, err
 }
